% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-input.R
\name{country_info}
\alias{country_info}
\title{Country code table}
\usage{
country_info()
}
\value{
A dataframe or tibble with the following columns:
\describe{
\item{\code{country_id}}{Country codes used by ORS}
\item{\code{name}}{English name of the country}
}
}
\description{
Information about ORS countries and their associated country codes.
Useful for the \code{avoid_countries} parameter in \code{\link{ors_params}}.
}
