% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois.R
\name{get_closest_pois}
\alias{get_closest_pois}
\title{Group a POI dataset}
\usage{
get_closest_pois(src, pois, n = NULL, radius = NULL)
}
\arguments{
\item{src}{\code{[sf]}

Source dataset containing point geometries that shall be routed from.}

\item{pois}{\code{[sf]}

Dataset that represents a list of points of interest to be routed to for each
row in the source dataset}

\item{n}{\code{[numeric]}

Maximum number of points of interest around each point in the source dataset
that shall be returned. The actual number might be lower depending on the
rows in the \code{pois} dataset and the remaining number of points if
\code{radius} is not \code{NULL}. If \code{NULL}, \code{radius} must be
provided.}

\item{radius}{\code{[numeric]}

Maximum distance of a point of interest around each point in the source
dataset. All returned points of interest lie within this distance to the
source points. If \code{NULL}, \code{n} must be provided.}
}
\value{
Returns \code{pois} with an added \code{.groups} column that links
each row to a row in the \code{src} dataset.
}
\description{
Groups a dataset containing points of interest based on their
proximity to a source dataset. Proximity can be defined through physical
distance (argument \code{n}) and/or a distance buffer (\code{radius}).
}
\examples{
\dontrun{
sample <- ors_sample(20)
pois <- get_osm_pois("Cologne", amenity = "hospital")

by_points <- get_closest_pois(sample, pois, n = 5)
by_buffer <- get_closest_pois(sample, pois, radius = 5000)
by_both <- get_closest_pois(sample_pois, n = 5, radius = 5000)
}

}
