% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_extract_boundaries}
\alias{get_extract_boundaries}
\title{Extract boundaries}
\usage{
get_extract_boundaries(instance = NULL, force = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{instance}{\code{[ors_instance]}

Object of an OpenRouteService instance that should be used for route
computations. It is recommended to use \code{\link{ors_instance}}
to set an instance globally. This argument should only be used if activating
an instance globally is not feasible.}

\item{force}{\code{[logical]}

If \code{TRUE}, extract must be identified and parsed. If
\code{FALSE}, the geometries will be read from the cache if possible.}

\item{verbose}{\code{[logical]}

If \code{TRUE}, prints a loading spinner.}

\item{...}{If \code{instance} is a local instance, this is ignored.
Otherwise, specifies further arguments for generating a random bbox
including: \code{dist}, the bbox radius in meters, and \code{seed}, a
random seed as in \code{\link{set.seed}}.}
}
\value{
An \code{sfc} object of the currently mounted extract boundaries.
}
\description{
Returns boundary geometries of the currently mounted extract
either from the local host or from a local cache.
}
