% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors-utils.R
\name{get_instance}
\alias{get_instance}
\alias{get_status}
\alias{get_profiles}
\alias{ors_ready}
\title{Utility functions}
\usage{
get_instance()

get_status(id = NULL)

get_profiles(id = NULL, force = TRUE)

ors_ready(id = NULL, force = TRUE, error = FALSE)
}
\arguments{
\item{id}{\code{[character]}

ID or name of a container or URL of a server that is to be checked. If
\code{NULL}, retrieves the ID from the current instance set by
\code{\link{ors_instance}}}

\item{force}{\code{[logical]}

If \code{TRUE}, function must query server. If \code{FALSE}, the information
will be read from the cache if possible.}

\item{error}{\code{[logical]}

If \code{TRUE}, gives out an error if the service is not ready.}
}
\value{
\code{get_instance} returns an object of class \code{ors_instance}.
\code{get_status} returns a list of information on the running service.
\code{ors_ready} returns a length-1 logical vector specifying if the service
is running. \code{get_profiles} returns a vector containing the active
profiles.
}
\description{
Utility functions to aid the setup of local instances.
\itemize{
\item \code{get_instance} checks for the existence of a mounted instance in
the current session and returns it.
\item \code{get_status} returns the status reported by the ORS server.
\item \code{ors_ready} checks if the mounted service is ready to use.
\item \code{get_profiles} is a wrapper around \code{get_status} that returns
the active profiles of the mounted service.
}
}
\seealso{
\code{\link{ors_instance}}
}
