% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois.R
\name{get_nearest_pois}
\alias{get_nearest_pois}
\title{Find the nearest points of interest from a local dataset}
\usage{
get_nearest_pois(
  source,
  pois,
  number_of_points = NULL,
  radius = NULL,
  crs = NA
)
}
\arguments{
\item{source}{Source dataset that represents point coordinates that are to
be routed from. The source dataset should be passed as a dataframe or plain
nested list with each row representing a x/y or lon/lat coordinate pair.}

\item{pois}{Dataset that represents a list of points of interest to be
routed to for each coordinate pair in the source dataset. The POI dataset
can either be passed as a dataframe or plain nested list with each row
representing a x/y or lon/lat coordinate pair, or, alternatively, as a
simple features object. If radius is not \code{NULL}, the POI dataset will be
converted to an sf object if it is not already.}

\item{number_of_points}{Integer scalar. Number of points to be returned.}

\item{radius}{Numeric scalar. Specifies the buffer size to select points of
interest by distance}

\item{crs}{Any object that is recognized by \code{\link[sf]{st_crs}}.
Specifies the coordinate notation of the source and pois dataset, if it they
are provided as simple coordinates.}
}
\value{
List of dataframes with each dataframe containing all points of
interest in a given type of proximity to the respective source point.
}
\description{
Returns points of interest in the proximity of the source
dataset. Unlike \code{\link{get_osm_pois}}, this function requires a local
dataset of points of interest.
}
\details{
The proximity can either be defined by the number of points to be selected, by a
distance buffer or by both. If both measures are defined, the function will select points
of interest within a certain radius first and will then select a given number of points
within that radius. The proximity type can be controlled by either passing or not passing
a value to \code{number_of_points} and \code{radius}.
}
\examples{

}
