% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{get_ors_geometry}
\alias{get_ors_geometry}
\alias{get_ors_summary}
\alias{get_ors_extras}
\alias{get_ors_attributes}
\alias{get_ors_waypoints_range}
\alias{get_ors_waypoints}
\alias{get_ors_warnings}
\alias{get_ors_features}
\alias{get_ors_alternatives}
\title{Low-level response helpers}
\usage{
get_ors_geometry(res, alt = 1L, as_coords = FALSE)

get_ors_summary(res, geometry = TRUE, ...)

get_ors_extras(res, which = NULL, alt = 1L)

get_ors_attributes(res, which = NULL, alt = 1L)

get_ors_waypoints_range(res, alt = 1L)

get_ors_waypoints(res, alt = 1)

get_ors_warnings(res)

get_ors_features(res, properties = TRUE)

get_ors_alternatives(res)
}
\arguments{
\item{res}{Parsed geojson response as returned by
\code{\link[=ors_inspect]{ors_inspect(..., as = "list")}}.}

\item{alt}{Alternative route to extract. Defaults to the recommended route.
Only relevant if alternative routes are computed.}

\item{as_coords}{Whether to return the geometry as an sf object or as
a dataframe.}

\item{geometry}{Whether to return the geometry in addition to each route
summary.}

\item{...}{Additional arguments passed to \code{get_ors_geometry}.}

\item{which}{Extra information (for \code{get_ors_extras}) or attribute
(for \code{get_ors_attributes}) to extract from the response. If \code{NULL},
extracts all extra information or attributes, respectively. Defaults to
\code{NULL}.}

\item{properties}{If \code{TRUE}, extracts feature properties.
If \code{FALSE}, only extracts properties (the nesting level before
properties). Only relevant if \code{res} is a geojson.}
}
\value{
\code{get_ors_geometry()} returns an object of class \code{sfc}
if \code{as_coords = FALSE}, otherwise a dataframe. If \code{res} is not
specified, returns an empty linestring. \code{get_ors_summary} returns an
object of class \code{sf} if \code{geometry = TRUE}, otherwise a dataframe.
\code{get_ors_extras} returns a 3-column matrix where the first column are
the start waypoints, the second column are the end waypoints and the third
column are the info codes. \code{get_ors_attributes} returns a dataframe
where each column is an attribute and each row is a segment.
\code{get_ors_waypoints_range} returns a numeric vector of waypoint indices.
\code{get_ors_waypoints} returns a dataframe where each row is a waypoint.
\code{get_ors_warnings} returns a 2-column dataframe containing the warning
code and message. \code{get_ors_features} returns a nested list.
\code{get_ors_alternatives} returns a single numeric value.
}
\description{
Low-level utilities to extract information from a routing response. These
functions are base-bones helpers that do not perform any type checks and
are mainly designed for internal use.

\itemize{
\item{\code{get_ors_geometry()} extracts the geometries from a
given response.}
\item{\code{get_ors_summary()} extracts the distance and duration of the
entire route.}
\item{\code{get_ors_extras()} extracts a matrix of waypoints and corresponding
extra info codes. See \code{\link{info_table}} to decode these values.}
\item{\code{get_ors_attributes()} extracts attribute values for each segment
of a route.}
\item{\code{get_ors_waypoints_range()} extracts waypoint intervals for each
route in the response.}
\item{\code{get_ors_waypoints()} extracts a dataframe of waypoints and
relevant metadata like distance, duration, and instructions.}
\item{\code{get_ors_warnings()} extracts all warnings returned along with
the response.}
\item{\code{get_ors_features()} extracts all features or properties of
a response.}
\item{\code{get_ors_alternatives()} returns the number of alternative
routes in a response.}
}
}
\examples{
\dontshow{httptest2::start_vignette("ors_inspect")}# retrieve an unformatted response
res <- ors_inspect(
  attributes = TRUE,
  extra_info = TRUE,
  navigation = TRUE,
  as = "list"
)

# These low-level utilities can be used to more conveniently extract
# information from complex ORS response structures.
get_ors_geometry(res)
get_ors_summary(res)
get_ors_extras(res, which = "steepness")
get_ors_attributes(res, which = "ascent")
get_ors_waypoints_range()
get_ors_waypoints()
get_ors_warnings()
get_ors_alternatives()
}
