% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors-utils.R
\name{get_profiles}
\alias{get_profiles}
\title{Get active ORS profiles}
\usage{
get_profiles(force = TRUE)
}
\arguments{
\item{force}{If \code{TRUE}, function must query the local host. If
\code{FALSE}, profiles will be read from the cache if possible.}
}
\description{
Returns a list of active profiles from the cache or local host.
}
