% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors-utils.R
\name{last_ors_conditions}
\alias{last_ors_conditions}
\title{Print ORS errors}
\usage{
last_ors_conditions(last = 1L)
}
\arguments{
\item{last}{Number of error lists that should be returned. \code{last = 2L},
for example, returns errors from the last two function calls.}
}
\description{
Return the error and warning messages that ORS returned in the
last \code{\link{get_route_lengths}} or \code{get_route_attributes} function calls.
}
