% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors_config.R
\name{ors_config}
\alias{ors_config}
\title{Change the ORS configuration}
\usage{
ors_config(
  instance,
  profiles = NULL,
  ...,
  matrix.maximum_routes = NULL,
  matrix.maximum_search_radius = NULL,
  isochrones.maximum_intervals = NULL,
  isochrones.maximum_locations = NULL,
  config = NULL,
  interactive = FALSE
)
}
\arguments{
\item{instance}{\code{[ors_instance]}

Object created by \code{\link{ors_instance}}.}

\item{profiles}{List of characters specifying the active profiles of an
instance. By default, can be one of the following: \code{car}, \code{hgv},
\code{bike-regular}, \code{bike-mountain}, \code{bike-road},
\code{bike-electric}, \code{walking}, \code{hiking}, \code{wheelchair}.}

\item{...}{Key-value pairs containing profile-specific routing options.
The argument name is interpreted as the profile whose configuration should be
changed. Pass \code{"default"} to change the default parameters. The argument
value needs to be a list of options to be changed.
Example: \code{car = c(maximum_distance = 100000000)}}

\item{matrix.maximum_routes}{Maximum number of routes that can be computed
at once. Defaults to 100.}

\item{matrix.maximum_search_radius}{Maximum allowed distance between a
coordinate and the nearest road. Defaults to 5000 m.}

\item{isochrones.maximum_intervals}{Maximum number of intervals for each
location. Defaults to 10.}

\item{isochrones.maximum_locations}{Maximum number of locations per request.
Defaults to 2.}

\item{config}{Parsed configuration file as contained in \code{ors_instance}
objects or path to a configuration file.}
}
\value{
Nested list of class \code{ors_instance}.
}
\description{
Activate profiles and make other changes to the OpenRouteService
configuration file. You can either make changes by passing a parsed
configuration file as contained in \code{ors_instance} objects or by
using one of the convenience arguments.
}
\seealso{
Other ORS setup functions: 
\code{\link{ors_change}()},
\code{\link{ors_extract}()},
\code{\link{ors_image}()},
\code{\link{ors_instance}()},
\code{\link{ors_remove}()},
\code{\link{ors_settings}()},
\code{\link{ors_start}()},
\code{\link{ors_up}()}
}
\concept{ORS setup functions}
