% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors_docker.R
\name{ors_down}
\alias{ors_down}
\title{Take down an ORS container}
\usage{
ors_down(instance)
}
\arguments{
\item{instance}{ORS instance object created by
\code{\link[ORSRouting]{ors_instance}}.}
}
\value{
Nested list of class \code{ors_instance}.
}
\description{
Stops and removes an OpenRouteService container. This is useful
if you need to rebuild a container, for example if profiles or extract were
changed.
}
\seealso{
Other ORS setup functions: 
\code{\link{ors_change}()},
\code{\link{ors_config}()},
\code{\link{ors_extract}()},
\code{\link{ors_instance}()},
\code{\link{ors_remove}()},
\code{\link{ors_settings}()},
\code{\link{ors_start}()},
\code{\link{ors_stop}()},
\code{\link{ors_up}()}
}
\concept{ORS setup functions}
