% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors_extract.R
\name{ors_extract}
\alias{ors_extract}
\title{Mount an extract to ORS}
\usage{
ors_extract(instance, place = NULL, provider = "geofabrik", file = NULL, ...)
}
\arguments{
\item{instance}{\code{[ors_instance]}

Object created by \code{\link{ors_instance}}.}

\item{place}{Place description of desired OpenStreetMap extract. This
argument is passed to \code{\link[osmextract]{oe_match}} which will try to
match it with an existing extract from a provider specified in \code{provider}.}

\item{provider}{OpenStreetMap extract provider that is supported by
\code{\link[osmextract]{oe_match}}}

\item{file}{Path or URL to an extract to be mounted.}

\item{...}{Further arguments passed to \code{\link[osmextract]{oe_match}}.}
}
\value{
Nested list of class \code{ors_instance}.
}
\description{
Add an extract to an existing OpenRouteService instance.
}
\seealso{
Other ORS setup functions: 
\code{\link{ors_change}()},
\code{\link{ors_config}()},
\code{\link{ors_image}()},
\code{\link{ors_instance}()},
\code{\link{ors_remove}()},
\code{\link{ors_settings}()},
\code{\link{ors_start}()},
\code{\link{ors_up}()}
}
\concept{ORS setup functions}
