% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors_docker.R
\name{ors_image}
\alias{ors_image}
\title{Manage the ORS image}
\usage{
ors_image(instance, tag = "latest", remove = FALSE, force = FALSE)
}
\arguments{
\item{instance}{\code{[ors_instance]}

Object created by \code{\link{ors_instance}}.}

\item{tag}{\code{[character]}

Docker reference tag. Defaults to \code{"latest"}.}

\item{remove}{\code{[logical]}

Whether to pull or remove an image. Cannot pull an image that already exists
and cannot remove an image that does not exist. Defaults to pulling.}

\item{force}{\code{[logical]}

Whether to force remove an image. This comes to play, e.g., when a container
is still running based on the image to be removed. Defaults to \code{FALSE}.
Ignored when \code{remove} is \code{FALSE}.}
}
\value{
Nested list of class \code{ors_instance}.
}
\description{
Pull or remove an OpenRouteService image
}
\seealso{
Other ORS setup functions: 
\code{\link{ors_change}()},
\code{\link{ors_config}()},
\code{\link{ors_extract}()},
\code{\link{ors_instance}()},
\code{\link{ors_remove}()},
\code{\link{ors_settings}()},
\code{\link{ors_start}()},
\code{\link{ors_up}()}
}
\concept{ORS setup functions}
