% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{ors_profile}
\alias{ors_profile}
\title{ORS profile}
\usage{
ors_profile(name = NULL, ..., template = TRUE)
}
\arguments{
\item{name}{\code{[character]}

Name of the ORS profile. If a length-2 vector is passed, the first
value is taken as the actual name and the second value is taken as the
title of the profile in the configuration file
(e.g. \code{c("driving-car", "car")}).}

\item{...}{Configuration parameters for the ORS profile. Must be key-value pairs. For
details, refer to the \href{https://giscience.github.io/openrouteservice/run-instance/configuration/ors/engine/profiles}{configuration reference}.}

\item{template}{\code{logical}

Whether to use a template if \code{name} is one of the base profiles
(see details).}
}
\value{
An object of class \code{ors_profile} that can be used to add
profiles to \code{\link{ors_instance}}
}
\description{
Construct the configuration for an ORS profile. The \code{ors_profile}
object created by this function can be used as input to the
\code{$add_profiles()} method of \code{\link[=ORSLocal]{ORS instances}}.
}
\details{
ORS defines a number of base profiles including:

\itemize{
\item{driving-car}
\item{driving-hgv}
\item{cycling-regular}
\item{cycling-road}
\item{cycling-electric}
\item{cycling-mountain}
\item{foot-walking}
\item{foot-hiking}
\item{wheelchair}
\item{public-transport}
}

All base profiles have a pre-defined template. Profiles other than
the base profiles can also be constructed with \code{ors_profile}, but
need some more sophisticated preparation (see e.g. Butzer 2017).
}
\examples{
# Create a car profile with pre-defined defaults
ors_profile("driving-car")

# Defaults can be modified
ors_profile("driving-car", elevation = FALSE)

# Create a walking profile with given encoder options
ors_profile("bike-regular", template = FALSE, encoder_options = list(turn_costs = FALSE))

}
\references{
Butzer, A.S. (2017). Erstellung eines Routing-Profils für
Feuerwehrfahrzeuge auf Basis von OpenStreetMap-Daten. Bachelor Thesis.
Ruprecht Karl University of Heidelberg.
}
