% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors-utils.R
\name{ors_ready}
\alias{ors_ready}
\title{Is ORS usable?}
\usage{
ors_ready(id = NULL, force = TRUE, error = FALSE)
}
\arguments{
\item{id}{\link{character}
ID or name of a container or URL of a server that is to be checked. If
\code{NULL}, retrieves the ID from the current instance set by
\code{\link{ors_instance}}}

\item{force}{\link{logical}
If \code{TRUE}, function must query server. If \code{FALSE}, the status will
be read from the cache if possible.}

\item{error}{\link{logical}
If \code{TRUE}, gives out an error if the service is not ready.}
}
\description{
States whether the ORS service is set up and ready to use.
}
