% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors_docker.R
\name{ors_remove}
\alias{ors_remove}
\title{Remove ORS instance}
\usage{
ors_remove(instance, ignore_image = TRUE)
}
\arguments{
\item{instance}{ORS instance object created by
\code{\link[ORSRouting]{ors_instance}}.}

\item{ignore_image}{If \code{TRUE}, does not remove the Docker image. This
can be useful, if the image is to be used for other purposes.}
}
\value{
Empty list of class \code{ors_instance}.
}
\description{
Fully remove any trace of an OpenRouteService instance. This is
a means of "uninstalling" OpenRouteService. It removes the container,
directory, and optionally the corresponding Docker image. It also disables
the instance object.
}
\seealso{
Other ORS setup functions: 
\code{\link{ors_change}()},
\code{\link{ors_config}()},
\code{\link{ors_down}()},
\code{\link{ors_extract}()},
\code{\link{ors_instance}()},
\code{\link{ors_settings}()},
\code{\link{ors_start}()},
\code{\link{ors_stop}()},
\code{\link{ors_up}()}
}
\concept{ORS setup functions}
