% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{ors_setup}
\alias{ors_setup}
\title{Local OpenRouteService setup}
\usage{
ors_setup(
  dir = "~",
  version = "master",
  overwrite = FALSE,
  place = NULL,
  provider = "geofabrik",
  profiles = NULL,
  ...,
  name = NULL,
  port = NULL,
  memory = NULL,
  auto_deletion = FALSE,
  wait = TRUE,
  tag = "latest",
  verbose = TRUE
)
}
\arguments{
\item{dir}{\code{[character]}

Custom OpenRouteService directory. If not specified, the directory
will be downloaded to the system's home directory. If a directory called
\code{"openrouteservice-{version}"} is present, the download will be skipped.
Ignored if \code{server} is not \code{NULL}.}

\item{version}{\code{[character]}

The OpenRouteService version to use. Can either be a version number (e.g.
6.7.0) or \code{"master"}. Ignored if \code{server} is not \code{NULL}.}

\item{overwrite}{\code{[logical]}

Whether to overwrite the current OpenRouteService directory
if it exists.}

\item{place}{Place description of desired OpenStreetMap extract. This
argument is passed to \code{\link[osmextract]{oe_match}} which will try to
match it with an existing extract from a provider specified in \code{provider}.}

\item{provider}{OpenStreetMap extract provider that is supported by
\code{\link[osmextract]{oe_match}}}

\item{profiles}{List of characters specifying the active profiles of an
instance. By default, can be one of the following: \code{car}, \code{hgv},
\code{bike-regular}, \code{bike-mountain}, \code{bike-road},
\code{bike-electric}, \code{walking}, \code{hiking}, \code{wheelchair}.}

\item{...}{Further arguments passed to \code{\link{ors_config}}.}

\item{name}{\code{[character]}

Name of the OpenRouteService container.}

\item{memory}{\code{[character]}

List of initial and maximum memory values. Initial memory is
the memory that is allocated to the Docker container from the start.
Maximum memory refers to the memory threshold that a Docker container cannot
exceed. If only a single value is passed, initial and maximum memory are
assumed to be identical.}

\item{auto_deletion}{\code{[logical]}

By default, OpenRouteService prevents all profiles other than \code{"car"}
from being built on first setup. If \code{FALSE}, disables this behavior.
Otherwise, all profiles other than car have to be enabled after the first
setup. Defaults to \code{TRUE}, because the OpenRouteService team recommends
building graphs for only the car profile in the initial setup.}

\item{wait}{\code{[logical]}

If \code{TRUE}, the function will not stop running after the container is
being started and will give out a notification as soon as the service is
ready. If \code{FALSE}, the function will start the container and then stop.
To check the server status, you can then call \code{\link{ors_ready}}.}

\item{tag}{\code{[character]}

Docker reference tag of the image to pull. Defaults to \code{"latest"}.}

\item{verbose}{\code{[logical]}

If \code{TRUE}, prints informative messages and spinners.}
}
\value{
Nested list of class \code{ors_instance}.
}
\description{
Performs a complete initial OpenRouteService server setup.
This function is a wrapper around \code{\link{ors_instance}},
\code{\link{ors_extract}}, \code{\link{ors_config}},
\code{\link{ors_settings}} and \code{\link{ors_up}} that simply serves
convenience. For a deeper understanding of what this function does, refer
to the corresponding function documentations.

After completing the setup, an instance object is returned and mounted to
the session. Functions that require OpenRouteService will automatically
access this object to determine where to send requests.
}
