% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors_docker.R
\name{ors_stop}
\alias{ors_stop}
\title{Stop an ORS container}
\usage{
ors_stop(instance)
}
\arguments{
\item{instance}{ORS instance object created by
\code{\link[ORSRouting]{ors_instance}}.}
}
\value{
Nested list of class \code{ors_instance}.
}
\description{
Stops a running OpenRouteService container. This can be used
over \code{ors_down}, if changes to the configuration were made that do not
require rebuilding the graphs (i.e. everything except profiles).
}
\seealso{
Other ORS setup functions: 
\code{\link{ors_change}()},
\code{\link{ors_config}()},
\code{\link{ors_down}()},
\code{\link{ors_extract}()},
\code{\link{ors_instance}()},
\code{\link{ors_remove}()},
\code{\link{ors_settings}()},
\code{\link{ors_start}()},
\code{\link{ors_up}()}
}
\concept{ORS setup functions}
