% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORSDockerInterface2.R
\name{ors_up}
\alias{ors_up}
\title{Start an ORS container}
\usage{
ors_up(instance, wait = TRUE, verbose = TRUE, distinct = TRUE)
}
\arguments{
\item{instance}{ORS instance object created by
\code{\link[ORSRouting]{ors_instance}}.}

\item{wait}{Logical. If \code{TRUE}, the function will not stop running
after the container is being started and will give out a notification as
soon as the service is ready. If \code{FALSE}, the function will start
the container and then stop. To check the server status, you can then
call \code{\link{ors_ready}}.}

\item{distinct}{Logical. If \code{TRUE}, adds a project name flag which
makes it possible to build multiple containers from the same image and
compose file. This is useful for running multiple instances at the same time.}
}
\value{
Nested list of class \code{ors_constructor}.
}
\description{
Build and start an OpenRouteService container.
}
\seealso{
Other ORS setup functions: 
\code{\link{ors_change}()},
\code{\link{ors_config}()},
\code{\link{ors_down}()},
\code{\link{ors_extract}()},
\code{\link{ors_instance}()},
\code{\link{ors_remove}()},
\code{\link{ors_settings}()},
\code{\link{ors_start}()},
\code{\link{ors_stop}()}
}
\concept{ORS setup functions}
