% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{ORSDocker}
\alias{ORSDocker}
\title{ORS Docker instance}
\description{
Creates a new ORS instance object that operates on a Docker instance of
OpenRouteService. This R6 class is typically constructed by
\code{\link{ors_instance}}. \code{ORSDocker} requires Docker to be
installed on the system and accessible by the current user. If these
requirements cannot be met, consider using \code{\link{ORSJar}} or
\code{\link{ORSWar}}.

For technical details on the setup of local ORS instances, refer to the
\href{https://giscience.github.io/openrouteservice/run-instance/running-with-docker}{Running with Docker documentation}.
For details on how to use \code{ORSDocker} objects, refer to the installation
vignette:

\preformatted{vignette("ors-installation", package = "rors")}
}
\details{
\code{ORSDocker} defines methods for all four steps in the ORS setup:
\itemize{
\item{\bold{Extract:} Download an extract and set it up for graph building.}
\item{\bold{Compose:} Change docker settings that control how the service is set up.}
\item{\bold{Configuration:} Change the way ORS computes routes.}
\item{\bold{Docker:} Send commands to docker to control the ORS container.}
}
}
\section{Docker commands}{


Local ORS instances are built using Docker containers. Initializing
\code{ORSLocal} downloads a \code{docker-compose.yml} file that tells
Docker how to build an ORS container. The Docker backend is pulled as
a Docker image (see
\href{https://hub.docker.com/r/openrouteservice/openrouteservice}{Dockerhub}).
\code{rors} communicates with Docker from within R to setup and manage ORS
containers.

On Unix systems, Docker requires superuser permissions. In Unix shells,
this is not problematic. However, R cannot communicate with Docker without
explicitly being granted superuser permissions or at least permissions to
access Docker. Thus, \code{ORSLocal} checks if the current user is the
superuser or if the current user has access to the Docker group. If not,
it aborts. For more details on how to manage Docker as a non-root user on
Linux, refer to the
\href{https://docs.docker.com/engine/install/linux-postinstall/#manage-docker-as-a-non-root-user}{Docker documentation}.
Note that this procedure grants root-level privileges to a user and can
negatively impact your system's security.
}

\section{Manual changes}{


\code{ORSLocal} provides a range of convenience methods to change the
setup of the ORS instance. All of these methods automatically read or write
their changes to disk. It is also possible to make direct changes to the
(parsed) configuration and compose files. In this case, reading and writing
also need to be done manually. To write changes to disk, run
\code{$update()}, e.g.:

\preformatted{ors$compose$parsed$services$`ors-app`$container_name <- "new-name-123"
ors$update()}

To read changes done manually to the files on disk, run
\code{$update("self")}.
}

\section{Configuration files}{


Contrary to what is written in the
\href{https://giscience.github.io/openrouteservice/run-instance/configuration/}{ORS documentation},
\code{ORSLocal} does not search for a pre-defined configuration file.
Instead, it explicitly creates a configuration file in the runtime directory.
Configurations are passed using an .env file but changes to the
configuration file are done using a .yml file due to easier reading and
writing within R. Accordingly, changes made to the .env file are
overwritten when \verb{$update("fs")} is called and cannot be read using
\verb{$update("self")}. All manual changes should be made to the .yml file.
Alternative configuration files can be specified by modifying the
\code{ORS_CONFIG_LOCATION} option in the compose file:

\preformatted{ors$compose$parsed$services$`ors-app`$environment$ORS_CONFIG_LOCATION <- "ors-config.yml"
ors$update()}
}

\examples{
\dontrun{
# Download ORS, start docker and jumpstart a default session
ors <- ors_instance(tempdir())

# Set a new extract file
ors$set_extract("Rutland")

# Allocate 100 MB of RAM
ors$set_ram(0.1)

# Add a routing profile
walking <- ors_profile("walking")
ors$add_profiles(walking)

# Set random port
ors$set_port()

# Change project name
ors$set_name("example-ors")

# Set up ORS
ors$up()

# If the setup fails or produces warnings, you can check
# using $show_logs()
ors$show_logs()

# Check if ORS container exists and is running
ors$is_built() # TRUE
ors$is_running() # TRUE

# Check if ORS is ready to use
ors$is_ready()

# You can stop and start a container
# This does not remove a container but only interrupts it
ors$stop()
ors$is_built() # TRUE
ors$is_running() # FALSE
ors$start()

# Take down container, e.g. to make configuration changes
# This removes a container so that you have to rebuild it
ors$down()
ors$is_built() # FALSE
ors$is_running() # FALSE

# Make changes to the configuration
ors$set_endpoints(matrix = list(maximum_routes = 1e+05)) # default is 2500

# Change default profile settings
default <- ors_profile(maximum_snapping_radius = -1)
ors$add_profiles(default)

# If we make manual changes to the configuration, we need to apply the
# changes explicitly
ors$config$parsed$ors$engine$profiles$car$elevation <- FALSE
ors$update() # writes the current object state to the disk

# If the compose or config files are changed on disk, the object can be
# refreshed
ors$update("self") # reads the disk state to the R object

bike <- ors_profile("bike-road")
ors$add_profiles(bike)

# Additionally, graphs are only re-built if we enable graph building.
# When changing the extract, this happens automatically, but we can also
# control graph building manually, e.g. for adding new profiles.
ors$set_graphbuilding(TRUE)

# Finally, start the container again to run the newly configured service
ors$up()
}
}
\section{Super classes}{
\code{\link[rors:ORSInstance]{rors::ORSInstance}} -> \code{\link[rors:ORSLocal]{rors::ORSLocal}} -> \code{ORSDocker}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{paths}}{List of relevant file paths for the ORS setup. Includes
the top directory, compose file, config file, and extract file.}

\item{\code{version}}{Version of the local ORS backend}

\item{\code{compose}}{Information of the compose file (\code{docker-compose.yml}).
The compose file holds various settings for the Docker setup. The most
important settings are included in this field:

\itemize{
\item{\code{ports}: A 2×2 matrix with Docker ports}
\item{\code{name}: Name of the Docker container}
\item{\code{memory}: List with memory information on total and free
system memory as well as initial and max memory allocated to the Docker
instance.}
\item{\code{image}: Version of the ORS image. \code{"latest"} refers
to the latest stable version. \code{"nightly"} refers to the devel
version.}
\item{\code{parsed}: Parsed compose file. When making changes to this
object, make sure to run \code{$update()} to apply the changes. For details,
refer to the \href{https://giscience.github.io/openrouteservice/run-instance/running-with-docker}{official reference}.}
}}

\item{\code{config}}{Information of the configuration file (\code{ors-config.yml}).
The config file holds various options about the ORS instance. This
field gives details about:

\itemize{
\item{profiles: A named vector of active routing profiles}
\item{parsed: Parsed configuration file. When making changes to
this obhect, make sure to run \code{$update()} to apply the changes.
For details, refer to the
\href{https://giscience.github.io/openrouteservice/run-instance/configuration/}{ORS reference}.}
}}

\item{\code{extract}}{Information on the extract file. Contains the name and
size of the selected extract file.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ORSDocker-new}{\code{ORSDocker$new()}}
\item \href{#method-ORSDocker-purge}{\code{ORSDocker$purge()}}
\item \href{#method-ORSDocker-set_name}{\code{ORSDocker$set_name()}}
\item \href{#method-ORSDocker-set_port}{\code{ORSDocker$set_port()}}
\item \href{#method-ORSDocker-set_memory}{\code{ORSDocker$set_memory()}}
\item \href{#method-ORSDocker-set_graphbuilding}{\code{ORSDocker$set_graphbuilding()}}
\item \href{#method-ORSDocker-set_image}{\code{ORSDocker$set_image()}}
\item \href{#method-ORSDocker-up}{\code{ORSDocker$up()}}
\item \href{#method-ORSDocker-down}{\code{ORSDocker$down()}}
\item \href{#method-ORSDocker-start}{\code{ORSDocker$start()}}
\item \href{#method-ORSDocker-stop}{\code{ORSDocker$stop()}}
\item \href{#method-ORSDocker-get_image}{\code{ORSDocker$get_image()}}
\item \href{#method-ORSDocker-get_container}{\code{ORSDocker$get_container()}}
\item \href{#method-ORSDocker-show_logs}{\code{ORSDocker$show_logs()}}
\item \href{#method-ORSDocker-is_built}{\code{ORSDocker$is_built()}}
\item \href{#method-ORSDocker-is_running}{\code{ORSDocker$is_running()}}
\item \href{#method-ORSDocker-is_init}{\code{ORSDocker$is_init()}}
\item \href{#method-ORSDocker-clone}{\code{ORSDocker$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rors" data-topic="ORSInstance" data-id="get_status"><a href='../../rors/html/ORSInstance.html#method-ORSInstance-get_status'><code>rors::ORSInstance$get_status()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rors" data-topic="ORSInstance" data-id="get_url"><a href='../../rors/html/ORSInstance.html#method-ORSInstance-get_url'><code>rors::ORSInstance$get_url()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rors" data-topic="ORSInstance" data-id="is_mounted"><a href='../../rors/html/ORSInstance.html#method-ORSInstance-is_mounted'><code>rors::ORSInstance$is_mounted()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rors" data-topic="ORSInstance" data-id="is_ready"><a href='../../rors/html/ORSInstance.html#method-ORSInstance-is_ready'><code>rors::ORSInstance$is_ready()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rors" data-topic="ORSLocal" data-id="add_profiles"><a href='../../rors/html/ORSLocal.html#method-ORSLocal-add_profiles'><code>rors::ORSLocal$add_profiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rors" data-topic="ORSLocal" data-id="report"><a href='../../rors/html/ORSLocal.html#method-ORSLocal-report'><code>rors::ORSLocal$report()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rors" data-topic="ORSLocal" data-id="rm_extract"><a href='../../rors/html/ORSLocal.html#method-ORSLocal-rm_extract'><code>rors::ORSLocal$rm_extract()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rors" data-topic="ORSLocal" data-id="rm_profiles"><a href='../../rors/html/ORSLocal.html#method-ORSLocal-rm_profiles'><code>rors::ORSLocal$rm_profiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rors" data-topic="ORSLocal" data-id="set_endpoints"><a href='../../rors/html/ORSLocal.html#method-ORSLocal-set_endpoints'><code>rors::ORSLocal$set_endpoints()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rors" data-topic="ORSLocal" data-id="set_extract"><a href='../../rors/html/ORSLocal.html#method-ORSLocal-set_extract'><code>rors::ORSLocal$set_extract()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rors" data-topic="ORSLocal" data-id="update"><a href='../../rors/html/ORSLocal.html#method-ORSLocal-update'><code>rors::ORSLocal$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-new"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-new}{}}}
\subsection{Method \code{new()}}{
Initialize the ORSDocker object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$new(
  dir = ".",
  version = NULL,
  overwrite = FALSE,
  dry = FALSE,
  verbose = TRUE,
  prompts = interactive(),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{\code{[character]}

Custom OpenRouteService directory. If not specified, the compose file
will be downloaded to the current working directory. If a directory called
\code{"openrouteservice-{version}"} is present, the download will be skipped.
Ignored if \code{server} is not \code{NULL}.}

\item{\code{version}}{\code{[character]}

The OpenRouteService version to use. Can either be a version number (e.g.
8.1.1) or \code{"master"}. Defaults to the most recent supported version.}

\item{\code{overwrite}}{\code{[logical]}

Whether to overwrite the current OpenRouteService directory if it exists.}

\item{\code{dry}}{\code{[logical]}

Whether to start a dry instance, i.e. initialize a docker instance without
requiring docker. This allows you to manipulate config, compose file,
and extract but does not allow you to interact with docker.}

\item{\code{verbose}}{\code{[logical]}

Level of verbosity. If \code{TRUE}, shows informative warnings and messages,
spinners, progress bars and system notifications.}

\item{\code{prompts}}{\code{[logical]}

Whether to ask for permission throughout the setup. Defaults to
\code{TRUE} in interactive sessions.}

\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-purge"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-purge}{}}}
\subsection{Method \code{purge()}}{
Purge ORS instance, i.e. take down container, (optionally) delete
image, delete ORS directory, and clean up R6 class.

This method can be useful for testing and writing reproducible
examples and can easily be used together with \code{\link{on.exit}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$purge(image = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image}}{\code{[logical]}

Whether to remove the docker image or keep it for other projects. The
default is \code{FALSE} to prevent accidentally breaking other projects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-set_name}{}}}
\subsection{Method \code{set_name()}}{
Set a name for the ORS container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$set_name(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name for the ORS container. If \code{NULL}, generates
a random name (\code{"ors-appXXXX"}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-set_port"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-set_port}{}}}
\subsection{Method \code{set_port()}}{
Set a port for the localhost of the ORS container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$set_port(port = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{port}}{\code{[numeric]}/\code{NULL}

Port to use for the container. Can be a vector of length 1 or 2.
The first port is for the API, the second port is optionally for
additional monitoring. If \code{NULL}, assigns a random port using
\code{\link[httpuv:randomPort]{randomPort()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-set_memory"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-set_memory}{}}}
\subsection{Method \code{set_memory()}}{
Set initial and max memory that the ORS container is allowed to use.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$set_memory(init = NULL, max = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{init}}{\code{[numeric/NULL]}

Initial memory in gigabytes. This can change if more memory is needed.
If not specified, uses \code{max}. If both are \code{NULL}, estimates
required memory.}

\item{\code{max}}{\code{[numeric/NULL]}

Maximum memory in gigabytes. The container is not allowed to use more
memory than this value. If not specified, uses \code{init}. If both are
\code{NULL}, estimates required memory.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-set_graphbuilding"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-set_graphbuilding}{}}}
\subsection{Method \code{set_graphbuilding()}}{
Graph building specifies whether routing graphs should be (re-)built.
Turning graph building on enables new profiles to be built or the
extract to be changed but significantly increases setup time. Turn
this off if you are changing configuration options that do not alter
routing graphs. Usually, you do not need to call this method
yourself because it is called automatically when a new extract is
set or when the setup is completed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$set_graphbuilding(mode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mode}}{\code{[logical]}

Whether to turn graph building on or off.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-set_image"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-set_image}{}}}
\subsection{Method \code{set_image()}}{
Set version of the ORS Docker image. The version should be compatible
with the compose version.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$set_image(version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{\code{[character]}

Version specification of the ORS image.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-up"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-up}{}}}
\subsection{Method \code{up()}}{
Create the ORS docker container and set up the ORS backend on a local
host.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$up(wait = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wait}}{\code{logical}

Whether to run a spinner and show a system notification when the setup
is completed. If \code{FALSE} releases the console after the Docker
container is created. You can then check the service status using
\code{$is_ready()}.}

\item{\code{...}}{Additional flags passed to the \code{docker up} command.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The exact command run by \code{$up()} is:

\code{docker compose -p [name] -f [compose file] up -d --no-build [...]}

The \code{-p} flag allows docker to run multiple ORS containers and keep
them separate. It uses the service name defined in the compose file.

If not found, \code{$up()} builds the underlying OpenRouteService docker
image specified by \code{version} during the initialization of
\code{ORSLocal}.

Usually in detach mode (\code{-d}), docker returns terminal control back
to the user. By default, \code{$up()} blocks the console, checks for
errors and notifies the user when the service setup has finished. This
behavior can be suppresed by setting \code{wait = FALSE}. The service
status can then be checked using \code{$is_ready()} or
\code{\link[=ors_ready]{ors_ready()}}. Container logs can be accessed by typing
\code{docker logs [name]} in the terminal.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-down"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-down}{}}}
\subsection{Method \code{down()}}{
Stop and remove the ORS docker container. Use this if you want to make
changes to a running ORS setup such as changing the extract or selected
profiles.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$down()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-start"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-start}{}}}
\subsection{Method \code{start()}}{
Start the ORS docker container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$start(wait = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wait}}{\code{logical}

Whether to run a spinner and show a system notification when the setup
is completed. If \code{FALSE} releases the console after the Docker
container is created. You can then check the service status using
\code{$is_ready()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-stop"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-stop}{}}}
\subsection{Method \code{stop()}}{
Stop the ORS docker container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-get_image"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-get_image}{}}}
\subsection{Method \code{get_image()}}{
Retrieve technical information on the docker image used.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$get_image()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-get_container"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-get_container}{}}}
\subsection{Method \code{get_container()}}{
Retrieve technical information on the docker container used.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$get_container()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-show_logs"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-show_logs}{}}}
\subsection{Method \code{show_logs()}}{
Show container logs as returned by \code{docker logs}. Useful for
debugging docker setups.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$show_logs(format = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{format}}{\code{[logical]}

If \code{TRUE}, includes ANSI colors and adds exdents. Otherwise,
trims ANSI colors. Disabling formatting increases performance, which
can be useful for larger logs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-is_built"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-is_built}{}}}
\subsection{Method \code{is_built()}}{
Checks if the ORS container is built. You can control this state
using \code{$up()} and \code{$down()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$is_built()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-is_running"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-is_running}{}}}
\subsection{Method \code{is_running()}}{
Checks if the ORS container is running. You can control this state
using \code{$start()} and \code{$stop()}. Check \code{$is_ready()} to see
if the ORS setup succeeded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$is_running()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-is_init"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-is_init}{}}}
\subsection{Method \code{is_init()}}{
Checks if ORS is initialized. ORS is initialized if it was built
for the first time. An initialized ORS instance has a subdirectory
called \code{"graphs"} that contains built graphs for at least one
routing profile. \code{$is_init()} therefore checks for the
existence of at least one sub-directory of \code{"graphs"}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$is_init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSDocker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ORSDocker-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDocker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
