% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ORSDockerInterface}
\alias{ORSDockerInterface}
\title{OpenRouteService Docker interface}
\description{
R6 class that interacts with the Docker daemon and is able to run basic commands
on the openrouteservice:latest image and the ors-app container
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{docker_running}}{Checks if the Docker daemon is running.}

\item{\code{image_built}}{Checks if the openrouteservice:latest image exists.}

\item{\code{container_exists}}{Checks if the container ors-app exists.}

\item{\code{container_running}}{Checks if the the container ors-app is running.}

\item{\code{service_ready}}{Checks if the container service is ready to use. If this field is \code{TRUE},
the service can be queried and used.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ORSDockerInterface$new()}}
\item \href{#method-image_up}{\code{ORSDockerInterface$image_up()}}
\item \href{#method-image_down}{\code{ORSDockerInterface$image_down()}}
\item \href{#method-start_container}{\code{ORSDockerInterface$start_container()}}
\item \href{#method-stop_container}{\code{ORSDockerInterface$stop_container()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes ORSDockerInterface, starts the Docker daemon and specifies the port.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDockerInterface$new(port = 8080)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{port}}{Integer scalar. Port that the server should run on.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-image_up"></a>}}
\if{latex}{\out{\hypertarget{method-image_up}{}}}
\subsection{Method \code{image_up()}}{
Builds the image, starts the container and issues a system notification when the
service is ready to be used.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDockerInterface$image_up()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-image_down"></a>}}
\if{latex}{\out{\hypertarget{method-image_down}{}}}
\subsection{Method \code{image_down()}}{
Deletes the image. Should only be used when the container does not exist.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDockerInterface$image_down()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_container"></a>}}
\if{latex}{\out{\hypertarget{method-start_container}{}}}
\subsection{Method \code{start_container()}}{
Starts the container and issues a system notification when the service is ready to
be used.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDockerInterface$start_container()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop_container"></a>}}
\if{latex}{\out{\hypertarget{method-stop_container}{}}}
\subsection{Method \code{stop_container()}}{
Stops the container.
Searches the OpenRouteService logs for the keyword 'error' and returns their
error messages.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSDockerInterface$stop_container()}\if{html}{\out{</div>}}
}

}
}
