% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ORSInstance}
\alias{ORSInstance}
\title{OpenRouteService backend control panel}
\description{
R6 class that acts as a setup wizard and control panel for the OpenRouteService backend service.
The class facilitates the setup of the Docker container and allows making changes to the setup from within R.
TODO: Interne Links zwischen den Docs
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dir}}{ORS directory, either passed as a parameter or automatically set after
calling \code{private$clone_ors_repo}}

\item{\code{extract}}{ORSExtract environment. See \code{?ORSRouting::ORSExtract}.}

\item{\code{config}}{ORSConfig environment. See \code{?ORSRouting::ORSConfig}.}

\item{\code{docker}}{ORSDockerInterface environment. See \code{?ORSRouting::ORSDockerInterface}.}

\item{\code{setup_settings}}{ORSSetupSettings environment. See \code{?ORSRouting::ORSSetupSettings}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ORSInstance$new()}}
\item \href{#method-finalize}{\code{ORSInstance$finalize()}}
\item \href{#method-get_config}{\code{ORSInstance$get_config()}}
\item \href{#method-init_docker}{\code{ORSInstance$init_docker()}}
\item \href{#method-get_setup_settings}{\code{ORSInstance$get_setup_settings()}}
\item \href{#method-set_ors_wd}{\code{ORSInstance$set_ors_wd()}}
\item \href{#method-init_setup}{\code{ORSInstance$init_setup()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize ORSInstance as well as OSMExtract and ORSConfig and ORSDockerInterface
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSInstance$new(dir = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{Custom ORS directory. If not specified, the directory will be downloaded
from the official GitHub repository.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
If necessary, stops the current ORS container when the class instance is removed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSInstance$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_config}{}}}
\subsection{Method \code{get_config()}}{
Initializes ORSConfig as an environment field. Call this if the config path changes
(e.g. after the initial ORS setup).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSInstance$get_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init_docker"></a>}}
\if{latex}{\out{\hypertarget{method-init_docker}{}}}
\subsection{Method \code{init_docker()}}{
Initializes DockerInstance as an environment field. This method starts docker and
delivers and interface to interact with docker.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSInstance$init_docker()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_setup_settings"></a>}}
\if{latex}{\out{\hypertarget{method-get_setup_settings}{}}}
\subsection{Method \code{get_setup_settings()}}{
Initializes ORSSetupSettings as an environment field and prepares the necessary
changes to the Dockerfile and docker-compose.yml
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSInstance$get_setup_settings(init_memory = NULL, max_memory = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{init_memory}}{Initial memory to be allocated to the docker container.}

\item{\code{max_memory}}{Maximum memory to be allocated to the docker container. The
container will start with the initial memory and increases the memory usage up to
the maximum memory if necessary.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_ors_wd"></a>}}
\if{latex}{\out{\hypertarget{method-set_ors_wd}{}}}
\subsection{Method \code{set_ors_wd()}}{
Tries to find the ORS directory and set it as the current working directory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSInstance$set_ors_wd(dir = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{Custom ORS directory. If not specified, the method will look inside the
default path and expects the default directory naming.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init_setup"></a>}}
\if{latex}{\out{\hypertarget{method-init_setup}{}}}
\subsection{Method \code{init_setup()}}{
Changes the necessary settings and configurations for the first startup, builds the
image and starts the container. This function should only be used when starting the service for
the first time. Changes after that should preferably be made manually.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSInstance$init_setup()}\if{html}{\out{</div>}}
}

}
}
