% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORSRouting.R
\docType{package}
\encoding{UTF-8}
\name{ORSRouting}
\alias{ORSRouting}
\title{ORSRouting: A package for routing distance computations using
OpenRouteservice}
\description{
The package allows the usage of OpenRouteService to calculate
the (shortest) routing distances of large datasets to certain points of
interest. It also provides functions to download or calculate nearby points
of interest. To process larger datasets, the package relies on a local
Docker setup of the OpenRouteService backend. The package facilitates the
setup by including a class that acts as a setup wizard and control panel.
}
\section{Exports}{

ORS setup and control panel:
\itemize{
\item \code{\link{ORSInstance}}
\item \code{\link{ORSExtract}}
\item \code{\link{ORSConfig}}
\item \code{\link{ORSSetupSettings}}
\item \code{\link{ORSDockerInterface}}
}

Points of interest data extraction and formatting:
\itemize{
\item \code{\link{get_osm_pois}}
\item \code{\link{get_nearest_pois}}
}

Route distance computation:
\itemize{
\item \code{\link{get_route_lengths}}
\item \code{\link{get_shortest_routes}}
}

Route inspection and summary:
\itemize{
\item \code{\link{inspect_route}}
\item \code{\link{summarize_route}}
}

Utility functions:
\itemize{
\item \code{\link{get_profiles}}
\item \code{\link{ors_ready}}
\item \code{\link{last_ors_conditions}}
\item \code{\link{grant_docker_privileges}}
}
}

\references{
This package is powered by OpenRouteService. For problems concerning their
service, refer to \url{https://github.com/GIScience/openrouteservice}
© openrouteservice.org by HeiGIT | Map data © OpenStreetMap contributors
}
