% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ORSSetupSettings}
\alias{ORSSetupSettings}
\title{R6 Docker setup control panel}
\description{
R6 class that controls docker-compose.yml and Dockerfile. Provides an interface to easily allocate
memory, switch graph building on or off and assign data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{compose}}{docker-compose.yml, parsed as a list. Blocks and items can be changed by assigning
values to them.}

\item{\code{init_memory}}{Initial memory to be allocated to the docker container. The container will
start with this amount of memory and will increase its memory usage if necessary.}

\item{\code{max_memory}}{Maximum memory to be allocated to the docker container. The container is not
allowed to exceed this memory limit.}

\item{\code{extract_path}}{Path where the OSM extract is saved.}

\item{\code{config_path}}{Path where the config file is saved.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{build_graphs}}{Switches graph building on or off. Graph building should be turned on when
the OSM extract changes or when no graphs exist yet. It should be turned off after building
graphs. If graph building is turned on after building the graphs, the service will rebuild all
graphs each time the image is rebuilt.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ORSSetupSettings$new()}}
\item \href{#method-allocate_memory}{\code{ORSSetupSettings$allocate_memory()}}
\item \href{#method-assign_data}{\code{ORSSetupSettings$assign_data()}}
\item \href{#method-save_settings}{\code{ORSSetupSettings$save_settings()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes the ORSSetupSettings class. Reads the docker-compose.yml, adjusts the Dockerfile
and allocates memory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSSetupSettings$new(
  extract_name,
  init_memory = NULL,
  max_memory = NULL,
  profiles = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{extract_name}}{File name of the extract}

\item{\code{init_memory}}{Initial memory to be allocated to the docker container.}

\item{\code{max_memory}}{Maximum memory to be allocated to the docker container. The
container will start with the initial memory and increases the memory usage up to
the maximum memory if necessary.}

\item{\code{profiles}}{Active profiles as set in the config file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allocate_memory"></a>}}
\if{latex}{\out{\hypertarget{method-allocate_memory}{}}}
\subsection{Method \code{allocate_memory()}}{
Specifies the amount of memory to be allocated. If only the memory limit is given,
the initial memory will be set to half that amount. If no memory is given, the method will
estimate the optimal amount of memory to be allocated. The memory is written to the compose file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSSetupSettings$allocate_memory(init = NULL, max = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{init}}{Initial memory to be allocated to the docker container.}

\item{\code{max}}{Maximum memory to be allocated to the docker container. The
container will start with the initial memory and increases the memory usage up to
the maximum memory if necessary.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assign_data"></a>}}
\if{latex}{\out{\hypertarget{method-assign_data}{}}}
\subsection{Method \code{assign_data()}}{
Writes the provided paths to the compose file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSSetupSettings$assign_data(mode = "build")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mode}}{Specifices whether the image is built for the first time or if the OSM extract is
being changed. \code{mode = build} inserts a \code{build} block that specifices the path to the extract
and config file. \code{mode = change} inserts a volume that overwrites the current graphs if graph
building is turned on.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_settings"></a>}}
\if{latex}{\out{\hypertarget{method-save_settings}{}}}
\subsection{Method \code{save_settings()}}{
Saves the setup changes by overwriting docker-compose.yml with all changed fields.
This should be run each time after changing any settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSSetupSettings$save_settings()}\if{html}{\out{</div>}}
}

}
}
