% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-input.R
\name{param_info}
\alias{param_info}
\title{Parameter info}
\usage{
param_info()
}
\value{
A dataframe or tibble with the following columns:
\describe{
\item{\code{name}}{Name of the parameter}
\item{\code{parent}}{Parent object of the parameter if it is nested, NA otherwise}
\item{\code{match}}{Does the parameter have a predefined list of values?}
\item{\code{scalar}}{Should the parameter be length-1?}
\item{\code{type}}{R type of the parameter value}
\item{\code{lim}}{Is parameter subject to numerical limits?}
\item{\code{allowed}}{Which profiles are required to use the parameter?}
\item{\code{poly}}{Should the parameter be formatted as a polygon?}
\item{\code{box}}{Does the parameter need boxing before POSTing?}
\item{\code{inspect}}{Is the parameter available for ors_inspect?}
\item{\code{pariwise}}{Is the parameter available for ors_pairwise?}
\item{\code{accessibility}}{Is the parameter available for ors_accessibility?}
}
}
\description{
Get information on the checks and preparations performed by
\code{\link{ors_params}} for additional ORS parameters.
}
