% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma.R
\docType{data}
\name{pharma}
\alias{pharma}
\title{Pharmacies in Rutland}
\format{
An sf tibble with 8 rows and 3 columns:
\describe{
\item{name}{Name of the pharmacy}
\item{dispensing}{Whether the pharmacy is able to provide prescription drugs}
\item{geometry}{sfc geometries}
}
}
\source{
\url{https://www.openstreetmap.org/},
\url{https://overpass-turbo.eu/}
}
\usage{
pharma
}
\description{
OpenStreetMap dataset of all pharmacies in the British county of Rutland.
}
\keyword{datasets}
