% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{plot_section}
\alias{plot_section}
\title{Plot route cross-sections}
\usage{
plot_section(
  x,
  feat = "avgspeed",
  dist = "distance",
  elev = "elevation",
  palette = NULL,
  scale_elevation = TRUE,
  smooth = FALSE,
  power = 1.5,
  span = 0.3,
  size = 1.5,
  xlab = dist,
  ylab = elev,
  scale_title = feat,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  ...
)
}
\arguments{
\item{x}{\code{[data.frame]}

A data.frame describing segments of a route. The data.frame
is expected to have multiple rows (representing the segments) and at least
three columns describing the elevation, distance, and an additional feature
to visualize. Preferably, this is a result of \code{ors_inspect}, but other
\code{sf} data.frames might work as well.}

\item{dist, elev, feat}{\code{[character]}

Column names of the distance, elevation and feature values inside \code{x}.
If \code{x} was computed using \code{elev_as_z = FALSE}, the elevation
data is derived from the Z variable.}

\item{palette}{\code{[character]}

Color palette to be used for plotting. Passed on to
\code{\link[ggplot2]{scale_fill_manual}} or
\code{\link[ggplot2]{scale_fill_gradientn}}, depending on the data type of
\code{feat}. Defaults to the Cividis colormap.}

\item{scale_elevation}{\code{[logical]}

Whether to scale the elevation axis based on the lowest elevation. If
\code{FALSE}, the y-axis is not scaled and fixed at sea level. Defaults to
\code{TRUE}.}

\item{smooth}{\code{[logical]}

Whether to smooth the elevation profile and the value
gradient. If \code{TRUE}, smoothes the elevation profile by first
interpolating the segment breaks and then fitting a LOESS model to
the elevation profile to create a more seamless figure. Interpolation is
performed using \code{\link{approx}} and LOESS smoothing using
\code{\link{loess}}.}

\item{power}{\code{[numeric]}

Factor by which to multiply the number of segment breaks.
A higher number equals a smoother blending of segments. A value of 1
indicates no interpolation (because \eqn{n \cdot 1 = 1}). Defaults to 1.5.}

\item{span}{\code{[numeric]}

Degree of LOESS smoothing. See \code{\link{loess}} for details.
A higher number equals rounder edges.}

\item{size}{\code{[numeric]}

Size of the line plot. Defaults to \code{1.5}.}

\item{xlab, ylab, scale_title, title, subtitle, caption}{\code{[character]}

Arguments to change label names, legend, title, subtitle and caption of the
ggplot object.}

\item{...}{Further arguments passed to
\code{\link[ggplot2]{scale_fill_manual}} or
\code{\link[ggplot2]{scale_fill_gradientn}} depending on the data type of
\code{feat}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the segments of a route as an elevation profile or
cross-section. A geographic cross-section describes a two-dimensional cut
through a route where the x-axis denotes the distance and the y-axis denotes
the elevation. The area below the elevation profile can be used to plot an
additional feature.
}
\examples{
\dontrun{
sample <- ors_sample(2)
route <- ors_inspect(sample[1, ], sample[2, ], extra_info = TRUE)

plot_section(route, feat = "waytype")

plot_section(route, feat = "steepness", scale_elevation = FALSE)

zissou_pal <- hcl.colors(11, palette = "Zissou 1")
plot_section(route, "steepness", palette = zissou_pal)

terrain_pal <- hcl.colors(10, palette = "Terrain")
plot_section(route, "elevation", palette = terrain_pal, scale_elevation = FALSE)
}

}
\references{
\url{https://giscience.github.io/openrouteservice-r/articles/openrouteservice.html}
}
