% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rors-package.R
\docType{package}
\encoding{UTF-8}
\name{rors-package}
\alias{rors}
\alias{rors-package}
\title{rors: Routing using 'OpenRouteService'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Flexible and opinionated interface to local and remote 'OpenRouteService' instances. The package enables qualitative and quantitative route computations, accessbility analyses, distance matrix calculations, street snapping and graph network exports. It is primarily aimed at processing larger datasets with controlled server-side limitations. R6 classes support the setup and and management of custom 'OpenRouteService' instances. The package functions take a tidy and graceful approach in querying and formatting the API output.
}
\section{Usage}{

To learn more about the usage of this package, look into the two
vignettes for this package:
\preformatted{
vignette("ors-installation", package = "rors")
vignette("ors-routing", package = "rors")
}
}

\section{Package options}{

The package relies on a number of package-level options that can be
customized:

\describe{
\item{\code{rors_setup_warn}}{If \code{TRUE}, emits warnings when a
log entry of type "WARN" is encountered during the ORS setup. This
concerns ORS setups using the \code{ORSLocal} class. Defaults to
\code{TRUE}.}

\item{\code{rors_echo}}{If \code{TRUE}, prints the results of a dry run
(using \code{\link[httr2]{req_dry_run}}) to the console each time a
request is sent to the mounted ORS server. Useful for debugging or
understanding what happens behind the scenes.}

\item{\code{rors_throttle_directions},
\code{rors_throttle_isochrones},
\code{rors_throttle_matrix},
\code{rors_throttle_snap}}{
Throttle specifications for each ORS service. \code{throttle}
indicates the number of allowed requests per minute that can be sent
to the server. The defaults can be retrieved from the
\href{https://openrouteservice.org/plans/}{ORS plans}. Throttle values
can either be a decimal number of (preferably) a fraction, e.g.
\code{40 / 60}. These options are useful for exploiting
rate limits of the collaborative plan or other ORS servers
that apply rate limits.}

\item{\code{rors_retries}}{Specifies the number of retries that are performed
before signaling an HTTP error. This option can be useful if the connection
to a server is unstable. Defaults to 3 retries.}
}
}

\references{
This package is powered by OpenRouteService. For problems concerning their
service, refer to \url{https://github.com/GIScience/openrouteservice}

© openrouteservice.org by HeiGIT | Map data © OpenStreetMap contributors
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JsLth/rors}
  \item \url{https://jslth.github.io/rors/}
  \item Report bugs at \url{https://github.com/JsLth/rors/issues}
}

}
\author{
\strong{Maintainer}: Jonas Lieth \email{jonas.lieth@gesis.org} [copyright holder]

}
\keyword{internal}
