% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors-sample.R
\name{get_extract_boundaries}
\alias{get_extract_boundaries}
\title{Extract boundaries}
\usage{
get_extract_boundaries(id = NULL, force = FALSE, verbose = TRUE)
}
\arguments{
\item{id}{\code{[character]}

ID or name of a container that contains the OSM extract file whose boundaries
should be extracted. If \code{NULL}, the default, retrieves the ID from the
current instance as set by \code{\link[ORSRouting]{ors_instance}}.}

\item{force}{\code{[logical]}

If \code{TRUE}, function must query local host. If
\code{FALSE}, the status will be read from the cache if possible.}

\item{verbose}{\code{[logical]}

If \code{TRUE}, prints a loading spinner.}
}
\value{
An \code{sfc} object of the currently mounted extract boundaries.
}
\description{
Returns boundary geometries of the currently mounted extract
either from the local host or from a local cache.
}
