% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois.R
\name{get_osm_pois}
\alias{get_osm_pois}
\title{Extract points of interest from OpenStreetMap}
\usage{
get_osm_pois(source, key, value, radius = 5000, crs = 4326)
}
\arguments{
\item{source}{Source dataset that represents point coordinates that are to be routed
from. The source dataset should be passed as a dataframe or plain nested list with
each row representing a x/y or lon/lat coordinate pair.}

\item{key}{Character scalar. OpenStreetMap Map Feature key (e.g. amenity, building)}

\item{value}{Character scalar. OpenStreetMap Map Feature value (e.g. hospital, hotel)}

\item{radius}{Numeric scalar. Radius in which POIs are to be searched for}

\item{crs}{Any object that is recognized by \code{sf::st_crs}. Coordinate reference system to determine
the coordinate notation of the source dataset}
}
\value{
List of dataframes that contain coordinate pairs of nearby points of interest
}
\description{
Makes Overpass requests to extract points of interest in the proximity of the source
dataset.
}
\examples{
pois <- query.osm.pois(datensatz.a, key = 'amenity', value = 'hospital', radius = 5000)
pois
# [[1]]
#          X        Y
# 1 6.960451 50.98735
# 2 6.924774 50.99160
# 3 6.975410 51.03098
#
# [[2]]
#          X        Y
# 1 6.957034 50.92491
# 2 7.042753 50.89777
# 3 7.050427 50.89228
# 4 6.966869 50.91035
#
# ...
}
