% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois.R
\name{get_osm_pois}
\alias{get_osm_pois}
\title{Extract points of interest from OpenStreetMap}
\usage{
get_osm_pois(
  source,
  ...,
  radius = 5000,
  crs = NA,
  trim = TRUE,
  timeout = NULL,
  as_sf = FALSE
)
}
\arguments{
\item{source}{Either (1) a source dataset as used for
\code{\link{ors_distances}}, (2) a character string of a place or
region, (3) a spatial polygon as an sf or sfc object or (4) any bbox format
supported by \code{\link[osmdata]{opq}}. The source argument is passed to
osmdata to derive a bounding box for the Overpass query.}

\item{...}{Can be used to pass key-value combinations of OSM map features.
Map features are passed as arguments: key = value. For a list of features,
refer to the
\href{https://wiki.openstreetmap.org/wiki/Map_features}{documentation}.}

\item{radius}{Numeric scalar. Specifies the buffer size if a source dataset
is passed (see details).}

\item{crs}{Any object that is recognized by \code{\link[sf]{st_crs}}.
Specifies the coordinate notation of the source dataset, if it is provided
(see details).}

\item{trim}{If TRUE and if a spatial polygon is passed as \code{source}, trim
the output data to the provided polygon shape.}

\item{timeout}{Timeout limit for the Overpass query. See
\code{\link[osmdata]{opq}}. If an sf object, a named vector, or a source
dataset is passed, this value can be estimated. For any other format, the
timeout is fixed to 100 seconds and may need adjustment for larger queries.}

\item{as_sf}{If TRUE, returns an sf dataframe containing point geometries.
If FALSE, returns a dataframe containing coordinates.}
}
\value{
If a source dataset is provided, either as sf or coordinate pairs, the
function will generate bounding boxes from buffers generated around each
coordinate pair in the dataframe. This results in a list of dataframes with
each element corresponding one coordinate pair in the source dataset.
If any other format is provided, the function will return a single dataframe
containing all points of interest within the provided area.
}
\description{
Makes Overpass requests to extract points of interest within
a specified area or list of areas.
}
\details{
The function tries to transform coordinate notations according to
the passed CRS. For vectors or matrices, this does not currently work. So
if you need to pass a vecorized bbox, make sure to pass geographic
coordinates, preferably EPSG:4326.
}
\examples{
\dontrun{
set.seed(123)
sample_a <- ors_sample(20)

pois_source <- get_osm_pois(sample_a, amenity = "hospital", radius = 5000, crs = 4326)

set.seed(123)
test_sf <- ors_sample(20, as_sf = TRUE)
test_poly <- sf::st_convex_hull(sf::st_union(test_sf))
get_osm_pois(test_poly, amenity = "post_box", building = "residential", trim = FALSE)

test_bbox <- as.numeric(sf::st_bbox(test_poly))
get_osm_pois(test_bbox, amenity = "marketplace", as_sf = TRUE)

test_place <- "Cologne"
get_osm_pois(test_place, landuse = "allotments", timeout = 200, as_sf = TRUE)
}
}
