% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors-utils.R
\name{get_profiles}
\alias{get_profiles}
\title{Get active ORS profiles}
\usage{
get_profiles(id = NULL, force = TRUE)
}
\arguments{
\item{id}{\link{character}
ID or name of a container or URL of a server that is to be checked. If
\code{NULL}, retrieves the ID from the current instance set by
\code{\link{ors_instance}}}

\item{force}{\code{[logical]}

If \code{TRUE}, function must query the local host. If
\code{FALSE}, profiles will be read from the cache if possible.}
}
\description{
Returns a list of active profiles from the cache or local host.
Requires OpenRouteService to be running.
}
