% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grant_docker_privileges}
\alias{grant_docker_privileges}
\title{Enable non-root docker access on Linux}
\usage{
grant_docker_privileges(run = TRUE)
}
\arguments{
\item{run}{If \code{FALSE}, the function will only return a logical vector
and will not change group memberships.}
}
\description{
Creates a docker group and adds the current user to it in order
to enable docker commands from within R. Doing this, either manually or by
using this function, is a requirement for using \code{\link{ORSInstance}}.
}
\details{
For details on what this function does and what security
implications it might have, refer to Docker's
\href{https://docs.docker.com/engine/install/linux-postinstall/}{post-installation guide}
}
