% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_table.R
\name{info_table}
\alias{info_table}
\title{Info tables}
\usage{
info_table(type)
}
\arguments{
\item{type}{The type of extra info to get information about. Must be one of
\code{"steepness"}, \code{"surface"}, \code{"waycategory"}, \code{"waytype"},
\code{"traildifficulty"}, \code{"roadaccessrestrictions"}, or
\code{"countryinfo"}.}
}
\value{
A tibble containing two two three columns containing the internal
code, a human-readable label and, if applicable, additional details.
}
\description{
Automatically generated info tables pulled from the
\href{https://giscience.github.io/openrouteservice/api-reference/endpoints/directions/extra-info/}{backend docs}.
These tables provide information about the possible extra info values that
can be retrieved when setting the \code{extra_info} argument in
\code{\link{ors_inspect}}.
}
\examples{
# get information about country codes used
info_table("country_list")
}
