% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-handling.R
\name{last_ors_conditions}
\alias{last_ors_conditions}
\title{Return ORS conditions}
\usage{
last_ors_conditions(last = 1L)
}
\arguments{
\item{last}{\code{[integer]}

Number of error lists that should be returned. \code{last = 2L},
for example, returns errors from the last two function calls.}
}
\description{
Return the error and warning messages that ORS returned in the
last \code{\link{ors_pairwise}} function call. Also works for
\code{\link{ors_shortest_distances}}.
}
