% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{ors_inspect}
\alias{ors_inspect}
\alias{summary.ors_route}
\title{Route inspection}
\usage{
ors_inspect(
  src,
  profile = NULL,
  level = c("waypoint", "step", "segment"),
  attributes = NULL,
  extra_info = NULL,
  elevation = TRUE,
  navigation = FALSE,
  alternative_routes = NULL,
  round_trip = NULL,
  as = c("tidy", "list", "string"),
  elev_as_z = FALSE,
  instance = NULL,
  ...,
  params = NULL
)

\method{summary}{ors_route}(object, ...)
}
\arguments{
\item{src}{\code{[sf]}

Source dataset containing at least two point geometries for which routes are
to be computed. If \code{round_trip} is specified, \code{source} must contain
a single coordinate pair from which a round trip is generated.}

\item{profile}{\code{[character]}

Means of transport as supported by OpenRouteService.
Defaults to the first profile in a call to \code{\link{get_profiles}}.
For \code{ors_shortest_distances}, \code{profile} can be a character
vector, for all other functions it needs to be a character scalar.
For details on all profiles, refer to the
\href{https://giscience.github.io/openrouteservice/technical-details/tag-filtering}{backend reference}.}

\item{level}{\code{[character]}

Level of route aggregation. Must be one of \code{waypoint}, \code{step} or
\code{segment}. See details.}

\item{extra_info}{\code{[character]}/\code{TRUE}

List of keywords that add extra information regarding each
linestring segment of the output. If \code{TRUE}, all values are included.
See details for more information.}

\item{elevation}{\code{[logical]}

If \code{TRUE}, elevation data is included in the output.}

\item{navigation}{\code{[logical]}

If \code{TRUE}, navigation information is included in the output, i.e.,
instructions and road exits. If \code{level == "segment"}, navigation data
is dropped from the output.}

\item{alternative_routes}{\code{[list]}

Named list that specifies options for alternative routes and accepts up to
three parameters. \code{target_count} is the maximum number of routes to
compute (including the recommended route). The output can contain less routes
than specified if no other alternatives can be computed from the \code{src}
coordinates. \code{share_factor} denotes the maximum share of identical paths
between routes. \code{weight_factor} is the maximum factor that a route can
deviate (i.e. be longer) from the original route. If specified, \code{src}
must contain exactly two coordinate pairs for which alternative routes
should be computed. If any alternatives can be computed given the specified
parameters, the alternative routes are added to the output by binding the
additional rows. In this case, a new column \code{alt} is added that informs
about the route type (\code{"recommended"} denotes the optimal route,
additional alternatives are termed \code{"alt1", "alt2", ...}). If
\code{NULL}, no alternative routes are computed.}

\item{round_trip}{\code{[list]}

Named list that specifies options for round trips and accepts up to three
parameters. \code{length} is the approximate length of the round trip.
\code{points} denotes the number of route points from which to derive a
round trip (the higher the rounder). \code{seed} controls the randomisation
of the route direction. If specified, \code{src} must contain a single
coordinate pair from which a round trip is to be generated. If \code{NULL},
no round trip is computed.}

\item{as}{\code{[character]}

How to format the output. If \code{"string"}, returns the entire JSON
response string. If \code{"list"}, returns a parsed JSON list. If
\code{"tidy"}, performs lots of data preparation to shape the response into a
rectangular shape. Defaults to \code{"tidy"}.}

\item{elev_as_z}{\code{[logical]}

If \code{TRUE}, elevation data is stored as z-values in the
geometry of the output \code{sf} dataframe. If \code{FALSE}, elevation is
stored as a distinct dataframe column. Ignored if \code{elevation = FALSE}.}

\item{instance}{\code{[ors_instance]}

OpenRouteService instance object created using \code{\link{ors_instance}}.
It is recommended to use \code{\link{ors_instance}} to set an instance
globally. This argument should be used if activating an instance
globally is not feasible or if subsequent requests using different instances
are sent.}

\item{...}{Additional arguments passed to the ORS API. Convenience way to
directly pass arguments of \code{\link{ors_params}}.}

\item{params}{\code{[ors_params]}

List of additional arguments passed to the ORS API. See
\code{\link{ors_params}} for details. If the same two arguments are passed
through \code{...} and \code{params}, the dot arguments take priority. Note
that the object passed through this argument needs to be built using the
same dataset and profile as in this function, otherwise an error is thrown.}

\item{object}{A route object of class \code{ors_inspect}.}
}
\value{
Returns an sf dataframe containing detailed sections of all routes
between the coordinates specified in \code{src}, aggregated according to
the level of aggregation stated in \code{level}. If \code{alternative_routes}
is specified and more than one route can be computed, adds a column
\code{alt} to the output that informs about the type of route alternative.
}
\description{
Calls the directions service once to get a closer look at route
characteristics and attributes. Produces a dataframe containing an elaborate
perspective on route sections and their contextual properties.
\code{ors_inspect} can also be used as a relatively low-level interface to
the directions service with the ability to return responses as parsed or
unparsed JSONs.

The summary function computes summary statistics and interval tables for
a given route.
}
\details{
OpenRouteService distinguishes between three
types of route aggregation: Segments, steps and waypoints. A \strong{segment}
is a single route between \code{src[i, ]} and \code{src[i + 1, ]}. A
\strong{step} is a route section as relevant for a navigation system. A
\strong{waypoint} is a straight connection between two geographical points
on a route.

Depending on the chosen level of aggregation, the output has to be adjusted
through interpolation and aggregation. Generally, \code{ors_inspect}
extracts information at the lowest level (waypoint) and then aggregates up.

\describe{
\item{If \code{level == "waypoint"}}{Extra information, elevation and
geometries are natively available at the waypoint level. Other data must be
interpolated or recycled. In particular, navigation information is recycled
from the step level. Distances are estimated based on the geometry length.
Durations are estimated based on the proportion of waypoint-level distances
to the route-level distance. From experience, estimated distances are rather
accurate while estimated durations and average speeds can deviate
considerably from the original information. Due to the estimation method
durations do not take into account elevation or road suitability.
Navigation information and street names do not perfectly overlap with
steps and segments. In these cases, the value with the highest overlap
is adopted causing some information loss. Information about the
corresponding step and segment index is attached to the output.}

\item{If \code{level == "step"}}{Navigation information and summary metrics
(distances, durations, average speed) are natively available at the step
level. Other data must be aggregated. Geometries are combined using
\code{\link[sf]{st_combine}}. Metric values are averaged using the mean.
Qualitative data is aggregated using the statistical mode, if multiple
unique values are found. Information about the corresponding segment index
is attached to the output.}

\item{If \code{level == "segment"}}{Attribute information (including
average speed) is natively available at the segment level. For waypoints
and steps, attributes are stored as an R attribute. For segments, this
information is attached to the output dataframe. For consistency, the
\code{avgspeed} attribute is always extracted and added to the output
Segments are aggregated using the same procedure as described above for
steps. Navigation information is always dropped as it is mostly meaningless
at an aggregation level this high.}
}

Extra information can be requested as additional context for each waypoint on
a route. Not all values are guaranteed to be available for all profiles.
If an extra info is not available, its column is added to the output and
filled with \code{NA}. Possible values include:

\itemize{
\item{\strong{steepness}: Ordered factor describing how steep a part of a
route is.}
\item{\strong{suitability}: Ordinal numeric describing how suitable a part
of a route is (1 - unsuitable; 10 - suitable).}
\item{\strong{surface}: Unordered factor describing the surface material
covering a part of a route.}
\item{\strong{waycategory}: Unordered factor describing special parts of
a route.}
\item{\strong{waytype}: Unordered factor containing different types of
roads.}
\item{\strong{tollways}: For \code{driving-*} profiles, specifies whether
a part of a
route is a tollway.}
\item{\strong{traildifficulty}: For walking and driving profiles, specifies
the OSM trail difficulty.}
\item{\strong{osmid}: For the wheelchair profile, contains the OSM IDs of
used ways.}
\item{\strong{roadaccessrestrictions}: Unordered factor describing road
access restriction.}
\item{\strong{countryinfo}: Nominal numeric containing country codes of a
part of a route.}
\item{\strong{green}: For walking profiles, describes the amount of green
on a route (1 - little; 10 - much).}
\item{\strong{noise}: For walking profiles, describes the amount of noise
on a route (1 - little; 10 - much).}
}
}
\examples{
\dontshow{httptest2::start_vignette("ors_inspect")}
# An instance must be mounted to run `ors_inspect`.
ors_instance(server = "pub")

# By default, only information about places, distances, durations,
# and elevations are returned
ors_inspect(pharma, "driving-car")

# Extra information can be controlled using the `extra_info`,
# `navigation`, and `attributes` arguments. You can either select
# specific information to be returned or pass TRUE to return all
# respective extra information
ors_inspect(
  pharma,
  navigation = TRUE,
  extra_info = TRUE,
  attributes = TRUE
)

# By default, `ors_inspect` partitions routes using the smallest possible
# division, i.e. waypoints. This comes at the cost of making some
# assumptions and interpolations. See details. To make partitions at
# a higher level, you can use the `level` argument.
ors_inspect(pharma, level = "step")
ors_inspect(pharma, level = "segment")

# If `attributes` are set, they are stored as actual R attributes.
# These attributes are general summaries that apply over the entire route.
insp <- ors_inspect(pharma, attributes = "detourfactor")
attr(insp, "detourfactor")

# By default, responses are tidied up and formatted in a workable dataframe.
# If you want to decide for yourself how to digest the retrieved data, you
# can choose to return the routes in an unformatted manner.
str(ors_inspect(pharma, as = "list"), max.level = 3)

# To create a round trip, only one point can be passed as `src`. This point
# is used as a basis to create a trip going from and back to the point.
ors_inspect(pharma[1, ], round_trip = list(length = 1000, points = 20))

# By default, only the optimal computed route is returned. To include
# possible alternatives as well, you can use `alternative_routes`.
# The following code computes 2 alternative routes where routes can use
# a maximum of 60\% overlapping ways.
ors_inspect(pharma, alternative_routes = list(target_count = 3, share_factor = 0.6))
}
\seealso{
\code{\link{ors_pairwise}},
\code{\link{plot_section}}
}
