% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{ors_inspect}
\alias{ors_inspect}
\alias{summary.ors_route}
\title{Route inspection}
\usage{
ors_inspect(
  src,
  profile = get_profiles(),
  level = c("waypoint", "step", "segment"),
  attributes = NULL,
  extra_info = NULL,
  elevation = TRUE,
  navigation = FALSE,
  alternative_routes = NULL,
  round_trip = NULL,
  as = c("tidy", "list", "string"),
  elev_as_z = FALSE,
  instance = NULL,
  ...
)

\method{summary}{ors_route}(object, ...)
}
\arguments{
\item{src}{\code{[sf]}

Source dataset containing at least two point geometries for which routes are
to be computed. If \code{round_trip} is specified, \code{source} must contain
a single coordinate pair from which a round trip is generated.}

\item{profile}{\code{[character]}

Character vector. Means of transport as supported by OpenRouteService. For a
list of active profiles, call \code{\link{get_profiles}}. For details on all
profiles, refer to the
\href{https://giscience.github.io/openrouteservice/documentation/Tag-Filtering.html}{documentation}.}

\item{level}{\code{[character]}

Level of route aggregation. Must be one of \code{waypoint}, \code{step} or
\code{segment}. See details.}

\item{attributes}{\code{[list]}

List of attributes that summarize route characteristics.
This includes three values: \code{avgspeed} states the average vehicle speed
along the route, \code{detourfactor} indicates how much the route deviates
from a straight line. \code{percentage} shows the share of a segment compared
to the entire route. If \code{TRUE}, all values are included.}

\item{extra_info}{List of keywords that add extra information regarding each
linestring segment of the output. If \code{TRUE}, all values are included.
See details for more information.}

\item{elevation}{\code{[logical]}

If \code{TRUE}, elevation data is included in the output.}

\item{navigation}{\code{[logical]}

If \code{TRUE}, navigation information is included in the output, i.e.,
instructions and road exits. If \code{level == "segment"}, navigation data
is dropped from the output.}

\item{alternative_routes}{\code{[list]}

Named list that specifies options for alternative routes and accepts up to
three parameters. \code{target_count} is the maximum number of routes to
compute (including the recommended route). Must be an integer between 1 and 3
and defaults to 1. The output can contain less routes than specified if no
other alternatives can be computed from the \code{src} coordinates.
\code{share_factor} denotes the maximum share of identical paths between
routes. \code{weight_factor} is the maximum factor that a route can deviate
(i.e. be longer) from the original route. If specified, and
\code{target_count} is larger than 1, the output is wrapped in a list of up
to three dataframes. If \code{NULL}, no alternative routes are computed.}

\item{round_trip}{\code{[list]}

Named list that specifies options for round trips and accepts up to three
parameters. \code{length} is the approximate length of the round trip.
\code{points} denotes the number of route points from which to derive a
round trip (the higher the rounder). \code{seed} controls the randomisation
of the route direction. If specified, \code{src} must contain a single
coordinate pair from which a round trip is to be generated. If \code{NULL},
no round trip is computed.}

\item{as}{\code{[character]}

How to format the output. If \code{"string"}, returns the entire JSON
response string. If \code{"list"}, returns a parsed JSON list. If
\code{"tidy"}, performs lots of data preparation to shape the response into a
tibble.}

\item{elev_as_z}{\code{[logical]}

If \code{TRUE}, elevation data is stored as z-values in the
geometry of the output \code{sf} dataframe. If \code{FALSE}, elevation is
stored as a distinct dataframe column. Ignored if \code{elevation = FALSE}.}

\item{instance}{\code{[ors_instance]}

Object of an OpenRouteService instance that should be used for route
computations. It is recommended to use \code{\link{ors_instance}}
to set an instance globally. This argument should only be used if activating
an instance globally is not feasible.}

\item{...}{Additional arguments passed to the ORS API. This includes all
options that modify the routing results. For details on each argument,
refer to the
\href{https://openrouteservice.org/dev/#/api-docs/v2/directions/{profile}/post}{API playground}
and
\href{https://giscience.github.io/openrouteservice/documentation/routing-options/Routing-Options.html}{documentation}
\describe{
\item{\strong{geometry_simplify}}{Whether geometry should be simplified.
Defaults to \code{FALSE}.}
\item{\strong{continue_straight}}{Whether to avoid u-turns and continue
straight if possible. If \code{FALSE},
allows the route to take u-turns.
Defaults to \code{FALSE}.}
\item{\strong{avoid_borders}}{Character vector specifying whether
to avoid, \code{all} borders, only
\code{controlled} ones or \code{none}. Only
available for \code{driving-*}.}
\item{\strong{avoid_countries}}{Numeric vector listing countries to avoid.
Each country is assigned a numeric value.
Refer to the \href{https://giscience.github.io/openrouteservice/documentation/routing-options/Country-List}{ORS country list}.
Only available for \code{driving-*}.}
\item{\strong{avoid_features}}{Character vector containing traffic features
to avoid. One or several of \code{highways},
\code{tollways}, \code{ferries}, \code{fords},
and \code{steps}.}
\item{\strong{avoid_polygons}}{\code{sf} or \code{sfc} object describing
areas to avoid.}
\item{\strong{restrictions}}{Restrictions to be applied for routing.
Routes that do not conform to these restrictions
are not considered. Can either be specifications
for heavy-goods vehicles (\code{axleload},
\code{hazmat}, \code{height}, \code{length},
\code{weight}, \code{width}), indicators
of wheelchair accessibility
(\code{maximum_incline},
\code{maximum_sloped_kerb}, \code{minimum_width},
\code{smoothness_type}, \code{surface_type},
\code{track_type}), or the maximum allowed
route steepness (\code{gradient}). Not
applicable for any other profile. Refer to the
\href{https://giscience.github.io/openrouteservice/documentation/routing-options/Routing-Options.html}{documentation}
for details on each parameter.}
\item{\strong{weightings}}{Weightings for route selection. Can be
preference multiplicators for walking profiles
(\code{green}, \code{quiet}) or
\code{steepness_difficulty} for cycling profiles.
Not applicable for any other profile.
Refer to the
\href{https://giscience.github.io/openrouteservice/documentation/routing-options/Routing-Options.html}{documentation}
for details on each parameter.}
\item{\strong{allow_unsuitable}}{Whether possibly unsuitable surfaces
should be included for wheelchair routing.
If \code{TRUE}, includes routes, even
if their surface quality might not be
suitable for wheelchair driving. Defaults
to \code{FALSE}.}
\item{\strong{surface_quality_known}}{Whether to enforce that the surface
quality of routes is considered.
If \code{FALSE}, routes of all surface
qualities are taken into account.
Defaults to \code{FALSE}}
\item{\strong{vehicle_type}}{The type of heavy-goods vehicle that is to be
assumed for profile \code{driving-hgv}. One of
\code{hgv}, \code{agricultural},
\code{delivery}, \code{forestry}, and
\code{goods}. Needed to set restrictions for
\code{driving-hgv}. Defaults to \code{hgv}.}
\item{\strong{preference}}{Specifies the route preference mode. One of
\code{recommended}, \code{fastest} or
\code{shortest}. Defaults to \code{recommended}.}
\item{\strong{radiuses}}{Maximum distance (in m) that road segments can be
snapped to. \code{radiuses = -1} represents an
unlimited radius. Defaults to the
\code{maximum_snapping_radius} in the configuration
file (350m if not changed).}
\item{\strong{maximum_speed}}{Maximum speed that routing vehicles are
allowed to drive. Not applied by default.}
}}

\item{object}{A route object of class \code{ors_inspect}.}
}
\value{
Returns an sf dataframe containing detailed sections of all routes
between the coordinates specified in \code{src}, aggregated according to
the level of aggregation stated in \code{level}. If \code{alternative_routes}
is specified, returns a list of sf dataframes instead with each element
containing a route alternative.

\code{ors_inspect} returns an sf data.frame containing the smallest
possible linestrings for a route along with additional information on each
segment. \code{ors_summary} returns an object of type \code{route_summary}
that contains information on distances, durations, speed, elevation, detour
factors as well as all available extra information for the requested route.
}
\description{
Calls the directions service once to get a closer look at route
characteristics and attributes. Produces a dataframe containing an elaborate
perspective on route sections and their contextual properties.
\code{ors_inspect} can also be used as a relatively low-level interface to
the directions service with the ability to return responses as parsed or
unparsed JSONs.

The summary function computes summary statistics and interval tables for
a given route.
}
\details{
OpenRouteService distinguishes between three
types of route aggregation: Segments, steps and waypoints. A segment is a
single route between \code{src[i, ]} and \code{src[i + 1, ]}. A step
is a route section as relevant for a navigation system. A waypoint is a
straight connection between two geographical points on a route.

Depending on the chosen level of aggregation, the output has to be adjusted
through interpolation and aggregation. For all levels below \code{"segment"},
ORS attributes are stored as R attributes and are not included in the
dataframe. For \code{"waypoint"}, distances and durations are derived from the
geometry lengths and do not take into account elevation (due to restrictions
in the s2 package). Extra information (using the \code{extra_info} argument)
and street names do not perfectly overlap with steps and segments. In these
cases, the value with the highest overlap is adopted causing some information
loss. Navigation information is dropped on \code{"segment"} level.

Extra information can be requested as additional context for each waypoint on
a route. Possible values include:
\describe{
\item{steepness}{Ordered factor describing how steep a part of a route is.}
\item{suitability}{Ordinal numeric describing how suitable a part of a
route is (1 - unsuitable; 10 - suitable).}
\item{surface}{Unordered factor describing the surface material covering a
part of a route.}
\item{waycategory}{Unordered factor describing special parts of a route.}
\item{waytype}{Unordered factor containing different types of roads.}
\item{tollways}{For \code{driving-*} profiles, specifies whether a part of a
route is a tollway.}
\item{traildifficulty}{For walking and driving profiles, specifies the OSM
trail difficulty.}
\item{osmid}{For the wheelchair profile, contains the OSM IDs of used ways.}
\item{roadaccessrestrictions}{Unordered factor describing road access
restriction.}
\item{countryinfo}{Nominal numeric containing country codes of a part of a
route.}
\item{green}{For walking profiles, describes the amount of green on a route
(1 - little; 10 - much).}
\item{noise}{For walking profiles, describes the amount of noise on a route
(1 - little; 10 - much).}
}
}
\examples{
\dontrun{
sample_source <- ors_sample(1)
sample_dest <- ors_sample(1)
profile <- get_profiles()[1]

# Basic inspection without extra information
insp <- ors_inspect(sample_source, sample_dest, profile)

# Advanced inspection with extra information
insp_adv <- ors_inspect(
  sample_source,
  sample_dest,
  profile,
  extra_info = TRUE
)

# Inspection of route elevation data
insp_elev <- ors_inspect(
  sample_source,
  sample_dest,
  profile,
  elevation = TRUE,
  elev_as_z = FALSE
)

# Inspection of route summary attributes
insp_attr <- ors_inspect(
  sample_source,
  sample_dest,
  profile,
  attributes = "detourfactor"
)
attr(insp_attr, "detourfactor")

# Altering the route by passing further arguments
insp_opts <- ors_inspect(
  sample_source,
  sample_dest,
  profile,
  continue_straight = TRUE,
  preference = "shortest",
  maximum_speed = 80
)

# Summarizing route specifics
route_summary <- ors_summary(sample_source, sample_dest, profile)
}
}
\seealso{
\code{\link{ors_pairwise}},
\code{\link{plot_section}}
}
