% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors_matrix.R
\name{ors_matrix}
\alias{ors_matrix}
\title{Routing distance matrix}
\usage{
ors_matrix(
  source,
  destination,
  profile = get_profiles(),
  units = c("m", "km", "mi"),
  proximity_type = c("distance", "duration"),
  instance = NULL
)
}
\arguments{
\item{source}{\code{[sf]}

Source dataset containing point geometries that shall be routed from.}

\item{destination}{\code{[sf]}

Destination dataset containing point geometries that shall be routed from.
The destination dataset follows the same format requirements as the source
dataset. For \code{ors_shortest_distances}, the destination argument can also
be a dataframe containing a grouping column specified by the \code{group}
argument that indicates which destinations refer to which row in the source
dataset (as returned by \code{\link{get_closest_pois}}). This is
recommended for large datasets because passing a plain sf dataframe routes
from each source point to each point in the entire destination dataset.}

\item{profile}{\code{[character]}

Character vector. Means of transport as supported by OpenRouteService. For a
list of active profiles, call \code{\link{get_profiles}}. For details on all
profiles, refer to the
\href{https://giscience.github.io/openrouteservice/documentation/Tag-Filtering.html}{documentation}.}

\item{units}{\code{[character]}

Distance unit for distance calculations (\code{"m"}, \code{"km"} or
\code{"mi"})}

\item{proximity_type}{\code{[character]}

Type of proximity that the calculations should be
based on. If \code{distance}, the shortest physical distance will be calculated
and if \code{duration}, the shortest temporal distance will be calculated.}

\item{instance}{\code{[ors_instance]}

Object of an OpenRouteService instance that should be used for route
computations. It is recommended to use \code{\link{ors_instance}}
to set an instance globally. This argument should only be used if activating
an instance globally is not feasible.}
}
\value{
If \code{length(proximity_type) == 1}, returns a
\code{nrow(source) * nrow(destination)} routing distance matrix. Otherwise,
returns a list containing two matrices accordingly.
}
\description{
Calls the matrix service and returns a routing distance matrix.
}
