% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors-utils.R
\name{ors_ready}
\alias{ors_ready}
\title{Is ORS usable?}
\usage{
ors_ready(force = TRUE, error = FALSE)
}
\arguments{
\item{force}{If \code{TRUE}, function must query local host. If
\code{FALSE}, the status will be read from the cache if possible.}
}
\description{
States whether the ORS service is set up and ready to use.
}
