% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors-sample.R
\name{ors_sample}
\alias{ors_sample}
\title{ORS dependent sampling}
\usage{
ors_sample(size, ..., as_sf = FALSE, force_new_extract = FALSE)
}
\arguments{
\item{size}{Number of points to be sampled}

\item{...}{Passed to \code{\link[sf]{st_sample}}, which passes it to
\code{\link[base]{sample}} or \code{\link[spatstat.core]{rThomas}}.}

\item{as_sf}{If \code{TRUE}, returns an \code{sfc} object, if \code{FALSE},
returns a dataframe containing coordinates.}

\item{force_new_extract}{If \code{TRUE}, forces the cached extract path to
be overwritten.}
}
\value{
\code{sfc} object containing the sampled \code{POINT} geometries
}
\description{
Sample point geometries within the boundaries of the running
ORS extract.
}
\details{
The function reads the extract file as an \code{sf} object.
Depending on the file, this can take a while. The unionized boundaries are
then cached making subsequent function calls a lot faster.
}
