% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{ors_sample}
\alias{ors_sample}
\title{ORS-based sampling}
\usage{
ors_sample(
  size,
  ...,
  force_new_extract = FALSE,
  instance = NULL,
  poly = NULL,
  verbose = TRUE
)
}
\arguments{
\item{size}{\code{[integer]}

Number of points to be sampled.}

\item{...}{Passed to \code{\link[sf]{st_sample}}.}

\item{force_new_extract}{\code{[logical]}

If \code{TRUE}, forces the cached extract path to be overwritten.}

\item{instance}{\code{[ors_instance]}

Object of an OpenRouteService instance that should be used for route
computations. It is recommended to use \code{\link{ors_instance}}
to set an instance globally. This argument should only be used if activating
an instance globally is not feasible.}

\item{poly}{\code{[sf/sfc]}

Boundary polygon used to sample points. If \code{NULL}, the default, boundary
polygons are extracted from the current instance as set by
\code{\link{ors_instance}}.}

\item{verbose}{\code{[logical]}

If \code{TRUE}, prints a loading spinner.}
}
\value{
\code{sfc} object containing the sampled \code{POINT} geometries
}
\description{
Sample point geometries within the boundaries of the running
ORS extract.

If the mounted instance is not local, the running ORS extract cannot be
identified. In this case, a random bounding box is generated. See
\code{\link{get_extract_boundaries}} for details.
}
\details{
The function vectortranslates the extract file to an \code{sf} object.
Depending on the file, this can take a while. The unionized boundaries are
then cached making subsequent function calls a lot faster.
}
