% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ors_settings.R
\name{ors_settings}
\alias{ors_settings}
\title{Change the ORS settings}
\usage{
ors_settings(
  instance,
  name = NULL,
  port = NULL,
  memory = NULL,
  auto_deletion = FALSE,
  graph_building = NULL
)
}
\arguments{
\item{instance}{\code{[ors_instance]}

Object created by \code{\link{ors_instance}}.}

\item{name}{\code{[character]}

Name of the OpenRouteService container.}

\item{memory}{\code{[character]}

List of initial and maximum memory values. Initial memory is
the memory that is allocated to the Docker container from the start.
Maximum memory refers to the memory threshold that a Docker container cannot
exceed. If only a single value is passed, initial and maximum memory are
assumed to be identical.}

\item{auto_deletion}{\code{[logical]}

By default, OpenRouteService prevents all profiles other than \code{"car"}
from being built on first setup. If \code{FALSE}, disables this behavior.
Otherwise, all profiles other than car have to be enabled after the first
setup. Defaults to \code{TRUE}, because the OpenRouteService team recommends
building graphs for only the car profile in the initial setup.}

\item{graph_building}{\code{[character]}

Mode of graph building to be applied. Graphs need to be built for each
profile and are required to route with a given profile. Possible values
include \code{"build"}, \code{"change"} and \code{NA}. \code{"build"} is used
when no graphs are built yet. \code{"change"} needs to be set before mounting
a new extract file to an existing setup. \code{NA} disables graph building.}

\item{ports}{\code{[numeric/character]}

Host port that the container should be running on. Passing \code{NULL}
is interpreted as "no change". Passing \code{NA} will assign a random port
using the \code{\link{httpuv}} package.}
}
\value{
Nested list of class \code{ors_instance}.
}
\description{
Allocate memory, manage graph building and set name and ports
of an OpenRouteService instance by making changes to the docker-compose file.
}
\seealso{
Other ORS setup functions: 
\code{\link{ors_change}()},
\code{\link{ors_config}()},
\code{\link{ors_extract}()},
\code{\link{ors_image}()},
\code{\link{ors_instance}()},
\code{\link{ors_remove}()},
\code{\link{ors_start}()},
\code{\link{ors_up}()}
}
\concept{ORS setup functions}
