% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ORSExtract}
\alias{ORSExtract}
\title{OpenRouteService OSM extract control panel}
\description{
R6 class to download, set or manage an OpenStreetMap extract
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Path to the extract. \code{Path} is not equal to \code{download_path} because the extract is copied
to docker/data after download.}

\item{\code{place}}{Name or coordinates of the place that is passed for the extract download.}

\item{\code{level}}{Hierarchical level of the extraxt to be matched.}

\item{\code{size}}{Size of the extract file in megabytes}

\item{\code{provider}}{Provider of the extract size.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ORSExtract$new()}}
\item \href{#method-get_extract}{\code{ORSExtract$get_extract()}}
\item \href{#method-set_extract}{\code{ORSExtract$set_extract()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes ORSExtract and looks for an existing extract in data path.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSExtract$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_extract"></a>}}
\if{latex}{\out{\hypertarget{method-get_extract}{}}}
\subsection{Method \code{get_extract()}}{
Downloads an OSM extract as pbf and moves it to the ORS data directory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSExtract$get_extract(place, download_path = tempdir(), level = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{place}}{Character scalar, sf or sfc object or length-2 numeric vector to be passed to
\code{osmextract::oe_match}. Represents a place that falls inside the coverage of defined extract
regions of the providers listed in \code{osmextract::oe_providers}. The geographic scale can be
adjusted by changing the parameter \code{level}. See details in \code{?osmextract::oe_match}.}

\item{\code{download_path}}{Character scalar. Path that the extract should be saved to.}

\item{\code{level}}{Integer scalar. If an sf or sfc object or a numeric vector is passed to
\code{place}, represents the hierarchical level of the extraxt to be matched. See details
in `?osmextract::oe_match.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_extract"></a>}}
\if{latex}{\out{\hypertarget{method-set_extract}{}}}
\subsection{Method \code{set_extract()}}{
Moves a given OSM extract to the ORS data directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSExtract$set_extract(extract_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{extract_path}}{Character scalar. Path to an OSM extract formatted as .pbf, .osm, .osm.gz
or .osm.zip.}
}
\if{html}{\out{</div>}}
}
}
}
