% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{ORSInstance}
\alias{ORSInstance}
\title{ORS instance}
\description{
This R6 class is the foundation for R6 classes \code{\link{ORSLocal}} and
\code{\link{ORSRemote}} and cannot be initialized. It defines methods to
check status and information of a remote ORS server and to mount itself to
the R session.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ORSInstance-is_ready}{\code{ORSInstance$is_ready()}}
\item \href{#method-ORSInstance-get_status}{\code{ORSInstance$get_status()}}
\item \href{#method-ORSInstance-is_mounted}{\code{ORSInstance$is_mounted()}}
\item \href{#method-ORSInstance-clone}{\code{ORSInstance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSInstance-is_ready"></a>}}
\if{latex}{\out{\hypertarget{method-ORSInstance-is_ready}{}}}
\subsection{Method \code{is_ready()}}{
Checks if ORS is ready to use. If \code{TRUE}, the \link{ORSRouting}
routing functions can be used.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSInstance$is_ready()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSInstance-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-ORSInstance-get_status}{}}}
\subsection{Method \code{get_status()}}{
Retrieves information about the ORS server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSInstance$get_status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSInstance-is_mounted"></a>}}
\if{latex}{\out{\hypertarget{method-ORSInstance-is_mounted}{}}}
\subsection{Method \code{is_mounted()}}{
Checks whether the instance is currently mounted to the session.
More technically, \code{$is_mounted} compares the MD5 hashes of its
R6 object and the currently mounted instance. Hence, an instance
only counts as mounted if it is exactly the same as the currently
mounted instance. An instance that was manually manipulated and has
not been updated with \code{$update()} thus is not considered
mounted.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSInstance$is_mounted()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ORSInstance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ORSInstance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ORSInstance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
