\name{ORSRouting-package}
\alias{ORSRouting-package}
\alias{ORSRouting}
\docType{package}
\title{
\packageTitle{ORSRouting}
}
\description{
\packageDescription{ORSRouting}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ORSRouting}
\packageIndices{ORSRouting}
The package allows the usage of OpenRouteService to calculate the
(shortest) routing distances of large datasets to certain points of
interest. It also provides functions to download or calculate nearby points
of interest. To process larger datasets, the package relies on a local
Docker setup of the OpenRouteService backend. The package facilitates the
setup by including a class that acts as a setup wizard and control panel.
}
\author{
\packageAuthor{ORSRouting}

Maintainer: \packageMaintainer{ORSRouting}
}
\references{
https://github.com/GIScience/openrouteservice
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
