\name{ORSRouting-package}
\alias{ORSRouting-package}
\alias{ORSRouting}
\docType{package}
\title{
\packageTitle{ORSRouting}
}
\description{
\packageDescription{ORSRouting}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ORSRouting}
\packageIndices{ORSRouting}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ORSRouting}

Maintainer: \packageMaintainer{ORSRouting}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
