% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORSRouting-package.R
\docType{package}
\encoding{UTF-8}
\name{ORSRouting-package}
\alias{ORSRouting}
\alias{ORSRouting-package}
\title{ORSRouting: Routing using OpenRouteService}
\description{
Flexible and opinionated interface to local and remote OpenRouteService instances. The package enables qualitative and quantitative route computations, accessbility analyses, distance matrix calculations, street snapping and graph network exports. It is primarily aimed at processing larger datasets with controlled server-side limitations. R6 classes support the setup and and management of custom OpenRouteService instances. The package functions take a tidy and graceful approach in querying and formatting the API output.
}
\section{Usage}{

To learn more about the usage of this package, look into the two
vignettes for this package:
\preformatted{
vignette("ors-installation", package = "ORSRouting")
vignette("ors-routing", package = "ORSROuting")
}
}

\references{
This package is powered by OpenRouteService. For problems concerning their
service, refer to \url{https://github.com/GIScience/openrouteservice}

© openrouteservice.org by HeiGIT | Map data © OpenStreetMap contributors
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JsLth/ORSRouting}
  \item \url{https://jslth.github.io/ORSRouting/}
  \item Report bugs at \url{https://github.com/JsLth/ORSRouting/issues}
}

}
\author{
\strong{Maintainer}: Lieth Jonas \email{jonas.lieth@gesis.org}

}
\keyword{internal}
