% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORSRouting-package.R
\docType{package}
\encoding{UTF-8}
\name{ORSRouting-package}
\alias{ORSRouting}
\alias{ORSRouting-package}
\title{ORSRouting: Batch routing distance computations using OpenRouteService}
\description{
Flexible interface to (custom) OpenRouteService instances. The package enables qualitative and quantitative route computations, accessibility analyses and distance matrix calculations using route distances (i.e., routing through real streets). It is primarily aimed at processing moderately large datasets and thus supports the setup and usage of custom OpenRouteService instances. However, it is also possible to use the public API given a valid key.
}
\section{Usage}{

To learn more about the usage of this package, look into the two
vignettes for this package:
\preformatted{
vignette("ors-installation", package = "ORSRouting")
vignette("ors-routing", package = "ORSROuting")
}
}

\references{
This package is powered by OpenRouteService. For problems concerning their
service, refer to \url{https://github.com/GIScience/openrouteservice}

© openrouteservice.org by HeiGIT | Map data © OpenStreetMap contributors
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JsLth/ORSRouting}
  \item Report bugs at \url{https://github.com/JsLth/ORSRouting/issues}
}

}
\author{
\strong{Maintainer}: Lieth Jonas \email{jonas.lieth@gesis.org}

}
\keyword{internal}
