% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route-inspection.R
\name{plot_section}
\alias{plot_section}
\title{Plot route cross-sections}
\usage{
plot_section(
  x,
  feat = "avgspeed",
  dist = "distance",
  elev = "elevation",
  palette = NULL,
  scale_elevation = TRUE,
  size = 1.5,
  xlab = dist,
  ylab = elev,
  scale_title = feat,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  ...
)
}
\arguments{
\item{x}{\code{[sf]}

An \code{sf} data.frame describing segments of a linestring. The data.frame
is expected to have multiple rows (representing the segments) and at least
three columns, \code{"elevation"}, \code{"distance"} and an additional
feature. Preferably, this is a result of \code{ors_inspect}, but other
\code{sf} data.frames might work as well.}

\item{dist, elev, feat}{\code{[character]}

Column names of the distance, elevation and feature values inside \code{x}.}

\item{palette}{\code{[character]}

Color palette to be used for plotting. Passed on to
\code{\link[ggplot2]{scale_fill_manual}} or
\code{\link[ggplot2]{scale_fill_gradientn}}, depending on the data type of
\code{feat}. Defaults to the Cividis colormap.}

\item{scale_elevation}{\code{[logical]}

Whether to scale the elevation axis based on the lowest elevation. If
\code{FALSE}, the y-axis is not scaled and fixed at sea level. Defaults to
\code{TRUE}.}

\item{size}{\code{[numeric]}

Size of the line plot. Defaults to \code{1.5}.}

\item{xlab, ylab, scale_title, title, subtitle, caption}{\code{[character]}

Arguments to change label names, legend, title, subtitle and caption of the
ggplot object.}

\item{...}{Further arguments passed to
\code{\link[ggplot2]{scale_fill_manual}} or
\code{\link[ggplot2]{scale_fill_gradientn}} depending on the data type of
\code{feat}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the segments of a route as an elevation profile or
cross-section. A geographic cross-section describes a two-dimensional cut
through a route where the x-axis denotes the distance and the y-axis denotes
the elevation. The area below the elevation profile can be used to plot an
additional feature.
}
\examples{
\dontrun{
sample <- ors_sample(2)
route <- ors_inspect(sample[1, ], sample[2, ], extra_info = TRUE)

plot_section(route, feat = "waytype")

plot_section(route, feat = "steepness", scale_elevation = FALSE)

zissou_pal <- hcl.colors(11, palette = "Zissou 1")
plot_section(route, "steepness", palette = zissou_pal)

terrain_pal <- hcl.colors(10, palette = "Terrain")
plot_section(route, "elevation", palette = terrain_pal, scale_elevation = FALSE)
}

}
\references{
\url{https://giscience.github.io/openrouteservice-r/articles/openrouteservice.html}
}
